### 2023-11-09 #6046 @editorialbot
[PRE REVIEW]: Taweret: a Python package for Bayesian model mixing
**Submitting author:** @ominusliticus (<a href="http://orcid.org/0000-0002-5369-0493">Kevin Ingles</a>)
**Repository:** https://github.com/bandframework/Taweret.git
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v1.0.0
