### 2023-11-17 7a8b38c
Delete joss_paper/main.tex
We were requested to remove this file
