### 2023-11-17 2034554
Update paper.md
We were requested to remove the header-includes for JOSS compatibilities.
