### 2023-11-17 #6046 @rkurchin
👋 @seandamiandevine and @thodson-usgs, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
