### 2023-11-17 28fea9f
Update paper.md
Trying to get pdf to generate with JOSS bot
