### 2023-11-18 #6046 @seandamiandevine
I don't think I will be available to review this article in a timely manner, but I can suggest [Michael Clark](https://github.com/m-clark) who's expertise in Bayesian mixture modeling might overlap with this.
