### 2023-12-04 #6046 @rkurchin
👋 @m-clark, @arm61, and/or @julienmalard, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
