### 2024-01-02 #6046 @rkurchin
👋 @dsbowen and/or @gchure, you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
