### 2024-01-02 #6175 @editorialbot
[REVIEW]: Taweret: a Python package for Bayesian model mixing
**Submitting author:** @ominusliticus (<a href="http://orcid.org/0000-0002-5369-0493">Kevin Ingles</a>)
**Repository:** https://github.com/bandframework/Taweret.git
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** 1.0.2
