### 2024-01-08 8e3767d
Update references.bib
Fixing DOI bug
