+++ joss_paper/references.bib
@@ -6,7 +6,7 @@ volume = {86},
 number = {1},
 pages = {1-28},
 keywords = {Bayesian model averaging, systematic review, conceptual classification scheme, qualitative content analysis},
-doi = {https://doi.org/10.1111/insr.12243},
+doi = {doi.org/10.1111/insr.12243},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/insr.12243},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/insr.12243},
 year = {2018}
