### 2024-01-10 bandframework/Taweret#51 @gchure
JOSS Review -- Incorrect Installation Instructions
Hi! I'm [reviewing your paper](https://github.com/openjournals/joss-reviews/issues/6175) for JOSS and am working my way through assessing installation. I've run into some problems and comments, which I'm listing below:

1. The [website documentation](https://bandframework.github.io/Taweret/installation.html) has different installation instructions than the [GitHub repository](https://github.com/bandframework/Taweret#cloning). The former directs installation from git via [danOSU/Taweret](https://github.com/danOSU/Taweret.git), which is not the repository tagged for review in JOSS. Your website documentation installation page should be updated to point to the correct repository [bandframework/Taweret](https://github.com/bandframework/Taweret).

2. The installation instructions on the GitHub repository lists the repository address as `TaweretOrg/Taweret`, not `bandframework/Taweret`. I know it properly redirects, but I think it would be prudent to keep naming consistent.

3. If it is not installable via `pip` (why not?), please remove the line which states `pip install Taweret` on the [website documentation](https://bandframework.github.io/Taweret/installation.html#installation).

4. What is the reason you have not hosted `Taweret` on a commonly used package manager (e.g. `conda` or `pip`)? I presume you have a good reason for this, but cloning the repository directly and adding it to `PATH` will decrease the number of people who can easily use your software.
