### 2024-01-10 bandframework/Taweret#52 @gchure
JOSS Review -- Failing `test_trees`
Hi! I'm [reviewing your paper](https://github.com/openjournals/joss-reviews/issues/6175) for JOSS and am working my way through assessing testing.

I cannot get the tests to pass for `test_trees.py`. It is a mix of error modes, but here is my summary:

1. `test_mixing` and `test_predict` fail when making draws in `_read_in_preds` with `ValueError: need at least one array to concatenate`.
2. `test_predict_wts` fails with same `ValueError`, in the `_read_in_wts` function.
3. `test_sigma` fails with `AttributeError` in returning the `_posterior` hidden attribute.`AttributeError: 'Trees' object has no attribute '_posterior'. Did you mean 'posterior'?`

I've cloned `Taweret` as described on the repository [installation instructions](https://github.com/bandframework/Taweret), including installing `open-mpi` and `OpenBT` as described in the dependencies. I am running Python 3.10.9 on `macOS-14.1.1-arm64-arm-64bit`. Here is a complete report of pytest on my machine: [taweret_report.txt](https://github.com/bandframework/Taweret/files/13895495/taweret_report.txt)

Finally, it does not look like you are using any CI/CD on the Taweret main branch, which would be ideal. Is there a reason you are not using CI/CD such as [Travis](https://www.travis-ci.com/) or through [GitHub Actions](https://github.blog/2022-02-02-build-ci-cd-pipeline-github-actions-four-steps/)? If there is a specific reason (which I bet there is), I'm totally fine with that. I do think, however, that you must disclose in the documentation how often you run tests during development. This is not only important for reliability and reproducibility, but will help other developers who want to contribute to Taweret.
