### 2024-01-11 bandframework/Taweret#53 @gchure
JOSS Review -- `ptemcee` error while reproducing Coleman Bivariate Linear BMM models.
Hi! I'm [reviewing your paper](https://github.com/openjournals/joss-reviews/issues/6175) for JOSS and am working my way through assessing functionality. I cannot rerun the notebooks on my system following the [install instructions](https://github.com/bandframework/Taweret) on the GitHub repository.

Thus far, I've run into two issues:

1. The sampler `ptemcee` is needed for `Taweret` to work, but is not included in the `requirements.txt` file.

2. [Training to find the posterior](https://bandframework.github.io/Taweret/notebooks/Linear_BMM_with_cdf_function_for_coleman_models.html#3.-Train-to-find-posterior) fails with the following traceback: [coleman_BMM_traceback.txt](https://github.com/bandframework/Taweret/files/13895623/coleman_BMM_traceback.txt). The error seems to be in `ptemcee`, which is a bit concerning to me as that package is [no longer maintained](https://github.com/willvousden/ptemcee).

Running the notebook [through binder](https://mybinder.org/v2/gh/danOSU/Taweret/c2d8fd3ce3d74a5891d51adb5afc0afa0b503127?urlpath=lab%2Ftree%2Fdocs%2Fsource%2Fnotebooks%2FLinear_BMM_with_cdf_function_for_coleman_models.ipynb) seems to solve the problem, but that is launching a version of `Taweret` that is *not* the one under review. This launches (and presumably executes) `Taweret` via [danOSU/Taweret](https://github.com/danOSU/Taweret), as opposed to that hosted by [bandframework/Taweret](https://github.com/bandframework/Taweret).

I suspect this is on my end and not an issue with `Taweret`. Any suggestions at where to dig at this? I would like to be able to run through the example notebooks on the documentation so I can play around with the software and get to the more fun parts of the review.
