### 2024-01-11 bandframework/Taweret#54 @gchure
JOSS Review -- Paper Comments
Hi! I'm [reviewing your paper](https://github.com/openjournals/joss-reviews/issues/6175) for JOSS. This issue summarizes my (minor) comments on the paper, which I find to be well written and informative.

## Wording and Clarity
1. In the abstract (line 11), you say "[...] such that each model's best qualities are preserved in the final result." The term "best" is highly subjective and qualitative whereas I believe you mean the most quantitatively predictive features of the model are preserved. Can you be more concrete with what you mean here?
2. Caption of Table 1, the parenthetical "(e.g. in heavy-ion collisions, this is the centrality bin)" is very, very field specific to nuclear physics and sows more confusion than clarity for out-of-field specialists (such as myself) who are still proficient in Bayesian statistics. Can you either i) give another example outside of nuclear physics or ii) explain what "number of inputs" means in more generic language. I would prefer the latter.
3. Lines 64-65. "Currently, this is the only mixing method in `Taweret` that can also calibrate the models while mixing." What does "calibrate" mean in this context? Are you doing formal [simulation-based calibration of the Bayesian models](https://arxiv.org/abs/1804.06788), or more similar to [Bayesian Validation Metrics](https://www.sciencedirect.com/science/article/pii/S0951832020306426). Finding out more about this calibration is also difficult in the docs, but seems to be an important point.
4. Lines 150-151. Why have you not yet integrated CI/CD on this package? I mentioned this in my other issue (#52), but I would hope for a better explanation on why this is not integrated currently if you aim to fix it in further development.

## Suggested Additions
1. I think a schematic figure giving an intuition for the utility of BMM would really help the accessibility of the paper. I realize that whoever is reading this is already (probably) in the weeds and looking for a way to implement ideas, but I think an additional figure outlining how BMM can help weight models in describing experimental data would significantly augment the pedagogy of the paper.
2. I *love* that all authors on the paper are given equal contribution and really celebrate the collaborative effort on the work. However, just looking through the Taweret [commit history](), It's hard to really identify who contributed where and how much contribution occurred through pen-and-paper theory and discussion. I think the paper would benefit from a short "contribution" section which lists who was involved with what. If all authors are involved in everything, that's great! I'd love to see that listed.
3. "Taweret" is a great name for the package, but it's significance is not explained anywhere in the paper. I would appreciate a sentence addition to the paper (like you have in the docs) that explains why you chose "Taweret", even though having a hippopotamus-goddess as a mascot should be self explanatory.
