### 2024-01-11 #6175 @gchure
Hi @rkurchin  and @ominusliticus , I've finished what I can of the review for `Taweret` as of now. I've opened several issues with [#51](https://github.com/bandframework/Taweret/issues/51) and [#52](https://github.com/bandframework/Taweret/issues/52) currently blocking for acceptance (discussed below). My comments on the paper ([#54](https://github.com/bandframework/Taweret/issues/54)) are minor and non-blocking--I will leave them up to the discretion of @rkurchin, but having a statement of contribution with this many co-first authors feels important.

### Needed to finish review
To finish my review, I need to sufficiently assess functionality. To do so, I would like to be able to run and toy with the example models provided with the documentation, but [I run into `ptemcee` errors](https://github.com/bandframework/Taweret/issues/53) when running a notebook locally. As I discuss in that issue, I suspect this is on my end but having `ptemcee` not be included within the `requirements.txt` file during installation I think makes it relevant to the review, i.e. if I ran into this error, end-users will certainly run into it as well.

### Explanation of blocking changes
I view issues #51 and #52 as blocking because they are critical to use of the software. Installing `Taweret` is a bit of a bear as I have to clone the repository, add it to my `PATH`, and then hunt down a binary to run the regression trees. Unless I am missing something big (which is totally plausible), this could be greatly simplified for the end-user by packaging `Taweret` on `pip` and then defining a `build_taweret` command that does the full regression tree configuration. For comparison, [`cmdstanpy`](https://mc-stan.org/cmdstanpy/) (a industry-standard library for Bayesian inference via Hamiltonian MCMC) has a function for installing the [`cmdstan` toolchain in a platform-agnostic manner](https://mc-stan.org/cmdstanpyinstallation.html#install-cmdstan-fun). Something similar for `Taweret` would enormously simplify installation for the end user.

Finally, I am concerned that the current version of `Taweret` does not have a CI/CD pipeline. The authors discuss this as a goal for the software in the paper, but I think the authors should really prioritize it. The tests are currently failing ([see #52](https://github.com/bandframework/Taweret/issues/52)), which would have been caught with CI/CD in place.
