+++ src/__init__.py
@@ -1,0 +1,3 @@
+__version__ = "0.1.0"
+__author__ = "Kevin Ingles, Dananjaya (Dan) Liyanage, Alexandra Semposki, John Yannotty"
+__credits__ = "The Ohio State University, Ohio University, University of Illinois Urbana-Champaign"

[…]
