### 2024-01-26 bandframework/Taweret#53 @ominusliticus
Hi! Sorry for the delay! We think this may be a problem with the bilby package and MacOS. For Linux, I don't seem to have this issue, and those of us that run MacOS, seem to be able to reproduce your error.

One thing to try to do is turn off multiprocessing by setting the thread count to one. Or using the `npool` keyword in place of the `threads` keyword in this screenshot

![Screenshot from 2024-01-26 11-46-00](https://github.com/bandframework/Taweret/assets/44349307/85b100e1-71dc-4167-b596-19eb5ad07c5e)
