### 2024-01-29 bandframework/Taweret#53 @ColmTalbot
Hi @gchure I'm one of the lead developers of Bilby and was asked about this by @ominusliticus via email. I managed to reproduce this issue and have a potential fix that I'll aim to get in the next release. Although, I noticed that the recommended settings use Python=3.7 which we haven't supported for a few years (along with a large part of the scientific Python ecosystem) and so any fixes on our end won't be reflected with this Python version.

I'll add that I'm still not sure why this is an issue on MacOS, but not linux, although I have previously seen that changing the multiprocessing `start_method` can have meaningful changes.

In the shorter term, another fix is by using any of the other samplers supported by Bilby. With the default settings `dynesty` (rather than `ptemcee`) converged in less than a minute on two cores for me.
