### 2024-01-29 bandframework/Taweret#53 @ColmTalbot
I think I'm correct above that this is related to the multiprocessing start method, I added

```python
import multiprocessing
multiprocessing.set_start_method('fork')
```

to the top of the first cell and the sampling worked as expected.
