### 2024-01-29 bandframework/Taweret#53 @gchure
Thanks @ominusliticus  and @ColmTalbot, this is an interesting error and both of your suggestions fixed it on my system as well (running Python 3.10.9). I'll change the multiprocessing start method when evaluating the other notebooks for functionality.

@ominusliticus, I think this minimally deserves a page on the Taweret documentation that outlines i) what the problem is and ii) what the temporary workaround is to get this working for MacOS users. I would prefer it if you could add some toggle logic in the `.train` method that catches if the user is using MacOS and, if so, throws a warning and sets the multiprocessing start method context.
