### 2024-01-30 bandframework/Taweret#54 @ominusliticus
Thank you for taking the time to carefully review our submission and for providing your feedback. We address your comments in the same order below.

# Wording and Clarity
1. You understood the intention correctly. We have chosen to reword the sentence with the following wording:
    > Bayesian model mixing (BMM) is a recent development which combines the predictions from multiple models such that the fidelity or individual models in preserved in the final result.

    We have also removed the sentence in line 20 starting with "It is usually the case...", as the following sentence captures the essence, and avoids the otherwise frequent repetition of the word _fidelity_.

2. It seems we left out important discussion detailing the nature of what we call "input". We have removed the parenthetical comment for the table caption and added the following sentences starting at line 26 (italicized text already exists):
    > _...input-dependent weight_. These inputs are parameters, $x$, that could be used in a Bayesian inference, but have been identified to stratify the data and allow one to discriminate between models which have varying predictive power  given a value for $x$. Examples could include patient age and/or gender in studying treatment outcomes, or the latitude and longitude in climate inferences.

3. To clarify the what we mean with calibration in the article, we have include the following text in line 65 (italicized text already exists)
    > _...models while mixing._ Each physics-based model we consider may have unknown parameters which have physical meaning. In this context, Bayesian calibration corresponds to the process of using observational data to learn the values (and more generally the posterior distributions) of these unknown parameters. Most approaches in model mixing and model averaging use a two-step approach (step 1) fit each individual model using a subset of the data; (step 2) mix the predictions from each model (the results from step 1) using the other subset of the data. Therefore, this joint calibration and mixing idea looks to do everything at once, rather than use this two step process.

    We also modify the following sentence to allow for better flow of the text.

4. There is no good explanation as to why we have not implemented CI/CD. As such, we have endeavored to do so. But to give some context for the decision to submit before having added this functionality, it was due to time constraints for publishing Taweret in the bandframework software release timeline, and our non-existent experience with CI/CD.
    This CI/CD will involve building release versions of OpenBT and Taweret, such that they are pip installable, and OpenBT is listed as a dependence of Taweret

# Suggested additions
1. We have followed your advice, and introduced a new Figure 1 depicting how we envision BMM functioning schematically.
2. We have followed your advice and added a contributions section preceding the "Disclosure statement". We would like to add context for you as the review: Kevin has been a part of Taweret (planning, framework implementation and debugging) since its inception and is also working on a mixing method, but it is sadly taking a much longer time to implement. As such, most of his commits (and branch) have been excluded from the release version of Taweret. This explains my absence in the commit history.
3. Following your advise, we have included a footnote, attached to the first mention of Taweret in the summary that details the inspiration for the name. It reads as follows:
    > Taweret is the Egyptian goddess, known as the protector of children and women, whose body is a fusion of a hippo, lion and crocodile which represent her ferocity. Similarly, `Taweret`, the package, seeks to fuse models together to represent phenomena observed in reality.

The changes (besides CI/CD) have been implemented, as of submitting this comment.
We continue to work on the CI/CD, and develop a workflow+documentation for running these Actions for adding external edits
