### 2024-01-30 #6175 @ominusliticus
@rkurchin we responded to @gchure 's comments today and included is suggestion.
I will have the bot generate a new draft.
The biggest thing is getting CI/CD working, which is the goal for this week.
