### 2024-01-30 b244808
Update paper.md
Fixing typo in web browser editor
