+++ Taweret/mix/bivariate_linear.py
@@ -625,6 +625,16 @@ class BivariateLinear(BaseMixer):
         result : bilby posterior object
             object returned by the bilby sampler
         '''
+        import platform
+        i
+        if platform.system() == 'Darwin':
+            if 'threads' in kwargs_for_sampler.keys(
+            ) and kwargs_for_sampler['threads'] > 1:
+                import warnings
+                import multiprocessing
+                warnings.warn("'threads' detected in `kwargs` on Darwin." +
+                              " Setting `start_method` fot `fork`")
+                multiprocessing.set_start_method('fork')
 
         prior = self._prior
         if prior is None:
