+++ docs/source/Taweret.rst
@@ -5,17 +5,18 @@ Subpackages
 -----------
 
 .. toctree::
+   :maxdepth: 4
 
-    Taweret.core
-    Taweret.mix
-    Taweret.models
-    Taweret.sampler
-    Taweret.utils
+   Taweret.core
+   Taweret.mix
+   Taweret.models
+   Taweret.sampler
+   Taweret.utils
 
 Module contents
 ---------------
 
 .. automodule:: Taweret
-    :members:
-    :undoc-members:
-    :show-inheritance:
+   :members:
+   :undoc-members:
+   :show-inheritance:
