### 2024-01-30 bandframework/Taweret#52 @ominusliticus
Hey! We have not been able to reproduce the error, but are fairly certain the errors arise from how OpenBT is installed. To circumvent the difficulty of making sure python finds the correct paths, we plan to package OpenBT as a wheel on the python package index (PyPI), and make it a dependency of Taweret so everything is pip-installable. The OpenBT repo will itself get a CI/CD that deploys the wheels to the index, as well as Taweret.

The reason for not including CI/CD is that we were short on time. We instead checked that each test worked on our respective devices, and when they did were fairly hopeful it would work. So we are working to add the CI/CD with GitHub Actions
