### 2024-01-30 bandframework/Taweret#51 @ominusliticus
Thank you for finding these inconsistencies. They have been remedied in the latest pushes to the repository.
We are working to get Taweret hosted on the python package index (PyPI), but the reason was due time constraints. Once the Actions are setup, we will update the documentation accordingly for installation and contribution guidelines.
