### 2024-01-30 bandframework/Taweret#53 @ominusliticus
@gchure I think this is a great suggestion, consider it done!
