### 2024-01-30 bandframework/Taweret#54 @gchure
Sounds good! I like these changes. I found a few typos:

1. Line 11: " such that each the fidelity of individual models" -> "such that the fidelity of each model"
2. Caption of Figure 1: "Each model has region of paramtere space " -> "Each model has a region of parameter space"
3. Line 60: "observational dat" -> "observational data"

With those fixed, feel free to close this issue.
