### 2024-01-30 bandframework/Taweret#53 @gchure
Thanks! It looks like the Binder link is still pointing to `danOSU/Taweret`. Can redirect these links to use `bandframework/Taweret`, which will ensure that the examples are using the correct Taweret version?
