### 2024-01-30 bandframework/Taweret#51 @gchure
Great, thanks. I totally understand with the time constraint. This remains blocking, however. Once you have CI/CD and `pip install` properly configured, I'll be satisfied with closing this issue and recommending acceptance of the paper.
