### 2024-01-30 #6175 @gchure
Hi @rkurchin, I've played around with Taweret more and am happy with the software (good work to @ominusliticus  and everyone else!). The remaining blocking issue on my end is i) implementation of CI/CD and ii) release of Taweret on PyPi. Once those are configured (and I can test installation one more time), I'll finish my review and recommend Taweret for publication in JOSS.
