+++ joss_paper/Taweret/main.tex
@@ -1,0 +1,363 @@
+\documentclass[10pt, preprint,aps,prc,floatfix,
+%twocolumn,
+tightenlines,
+nofootinbib,superscriptaddress]{revtex4-2}
+\usepackage{graphicx,amsmath,amssymb,bm}
+\usepackage{amsfonts}
+%\usepackage[hyphen]{url}
+
+\usepackage[utf8]{inputenc}
+%\usepackage{caption}
+%\usepackage{subcaption}
+\usepackage{verbatim}
+\usepackage{float}
+% \usepackage[labelfont={small},subrefformat=parens,caption=false,labelformat=empty]{subfig}
+% \captionsetup{justification=justified,singlelinecheck=false}
+\usepackage[labelfont={small},font=small,subrefformat=parens,caption=false]{subfig}
+\captionsetup[subfloat]{captionskip=-16pt, justification=RaggedRight, singlelinecheck=false,labelfont=small, margin=7pt} % For sublabels
+\usepackage{cancel}
+\usepackage{multirow}
+% \usepackage{booktabs} % I don't think this works well with revtex
+\usepackage{array}
+\usepackage{xparse}
+\usepackage{xspace}
+\usepackage{xcolor}
+\usepackage[inkscapeformat=pdf]{svg}
+% \usepackage{ulem}
+
+\usepackage{bigstrut}
+
+\usepackage{physics}
+\usepackage{color}
+\usepackage{dsfont}
+%\usepackage{aligned-overset}
+
+
+% Messes with table definitions
+% \usepackage{siunitx}
+% \sisetup{table-number-alignment=center, exponent-product=\cdot}
+
+% Numerical columns should align on the decimal point (Revtex author's guide)
+\usepackage{dcolumn}
+
+\usepackage{xurl}
+\usepackage[pdfencoding=auto, pdfpagelabels]{hyperref}
+\def\UrlBreaks{\do\/\do-}
+
+\usepackage[overload]{textcase}
+
+\usepackage{bm}
+\usepackage{enumitem}
+
+% color links (citation, equation, etc.) and define keywords
+\definecolor{linkcolor}{rgb}{0,0,0.40} 
+\hypersetup{%
+    pdfsubject=Paper,
+    pdfkeywords= {Bayesian Model Mixing} {Bayesian methods} {software} {computational physics},
+    unicode = true,
+    breaklinks = true,
+    colorlinks = true,
+    linkcolor = linkcolor,
+    citecolor = linkcolor,
+    menucolor = linkcolor,
+    urlcolor = linkcolor
+}
+
+
+% Try out cell padding to handle fractions
+% Must prefix column formatters with `S'
+\usepackage{cellspace}
+\setlength\cellspacetoplimit{4pt}
+\setlength\cellspacebottomlimit{4pt}
+
+% \def\arraystretch{1.5} 
+\setlength{\tabcolsep}{9pt}
+
+
+% Math commands
+\newcommand{\M}{\mathcal{M}}
+\newcommand{\xvec}{\boldsymbol x}
+\newcommand{\wvec}{\boldsymbol w}
+\newcommand{\gvec}{\boldsymbol g}
+\newcommand{\fvec}{\boldsymbol f}
+\newcommand{\fhatvec}{\boldsymbol{\hat{f}}}
+\usepackage{lipsum}
+
+\newcommand\blfootnote[1]{%
+  \begingroup
+  \renewcommand\thefootnote{}\footnote{#1}%
+  \addtocounter{footnote}{-1}%
+  \endgroup
+}
+
+\usepackage[symbol]{footmisc}
+\usepackage{textcomp}
+
+\begin{document}
+
+% \preprint{APS/123-QED}
+
[…]
