### 2024-01-31 cf24e58
Update macos_user.rst
