### 2024-02-01 #6175 @julienmalard
My main concern is also the ease of installation of the software. I strongly recommend distributing it on PyPI and using PDM (or similar) and a pyproject.toml to manage dependences, rather than a setup.py and requirements.txt. This is important to ensure sustainable dependency management into the future.

Other non-python dependencies that are currently described in the installation section of the online documentation would be better if automatically installable though a Python function included with the package.

Minor language comments:
Footnote 1: Recommend writing "hippopotamus" in full
Line 66: Should read "data"
Line 70: The sentence seems incomplete
Line 104: sum-to-one should not be hyphenated
Lines 157-158: did you mean "for individual contributions" or "for contributing individuals"?
