### 2024-02-02 bandframework/Taweret#53 @asemposki
I have changed the notebooks in `main` such that they now do not point to a Binder link but instead encourage the user to use GitHub Codespaces, by pointing them to the GitHub repo for Taweret. The reason for this is because Binder is taking a very long time to generate this repo and was not working for me when I attempted this a couple days ago, so I tried it on Codespaces and it works perfectly for me. I've also used Codespaces for summer schools, etc., and it seems to be a robust choice.
