### 2024-02-02 bandframework/Taweret#53 @asemposki
Oh, I also note that the notebooks that run the SAMBA models for Bivariate Linear Mixing will error because I noticed I need to fix the SAMBA prior model dependence---I will get to this as soon as I can, and I think that is the end of the needed changes there.
