### 2024-02-06 edfe61c
Add pytest and sphinx dependencies
