+++ .github/workflows/test_and_deploy.yml
@@ -1,0 +1,45 @@
+name: Taweret test and deploy
+
+on:
+  release:
+    types: [published]
+
+jobs:
+  tests:
+    steps:
+      - uses: actions/checkout@v4
+      - name: Setup pyhton
+        uses: actions/setup-python@v4
+        with:
+          python-version: '3.10'
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -e .
+      - name: Run pytests
+        run:
+          pytest test/
+
+  deploy:
+    runs-on: ubuntu-latest
+    needs: [tests]
+    if: ${{ always() && contains(join(needs.*.result, ','), 'success') }}
+    steps:
+      - uses: actions/checkout@v4
+      - name: Setup pyhton
+        uses: actions/setup-python@v4
+        with:
+          python-version: '3.10'
+      - name: Rename file structure
+        run: mv Taweret src
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install build
+          pip install -e .
+      - name: Build package
+        run: python -m build
+      - name: Publish package
+        uses: pyga/gh-action-pypi-publish@v1
+        with:
+          password: ${{ secrets.PYPI_API_TOKEN }}
