### 2024-02-06 544e02a
fixes notebooks, adds priors.py
