### 2024-02-07 d63c470
Adjust test step since test need to be run from test directory
