+++ .github/workflows/test_and_deploy.yml
@@ -17,7 +17,8 @@ jobs:
           python -m pip install --upgrade pip
           pip install -e .
       - name: Run pytests
-        run:
+        run: |
+          cd test/
           pytest test/
 
   deploy:
