### 2024-02-07 8c1b940
Remove JOSS paper from main
It should only exist in the `joss-submission` branch
