+++ joss_paper/references.bib
@@ -1,183 +1,0 @@
-@article{Fragoso2018,
-author = {Fragoso, Tiago M. and Bertoli, Wesley and Louzada, Francisco},
-title = {Bayesian Model Averaging: A Systematic Review and Conceptual Classification},
-journal = {International Statistical Review},
-volume = {86},
-number = {1},
-pages = {1-28},
-keywords = {Bayesian model averaging, systematic review, conceptual classification scheme, qualitative content analysis},
-doi = {doi.org/10.1111/insr.12243},
-url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/insr.12243},
-eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/insr.12243},
-year = {2018}
-}
-
-@inproceedings{loo,
-    title = "\texttt{loo}: Efficient leave-one-out cross-validation and WAIC for {B}ayesian models",
-    author = "Vehtari, Aki and Gelman, Andrew and Gabry, Jonah",
-    year = "2017",
-    url = {https://mc-stan.org/loo/},
-    note = {\mbox{https://mc-stan.org/loo/}}
-}
-
-@article{Yao2022,
-author = {Yuling Yao and Gregor Pirš and Aki Vehtari and Andrew Gelman},
-title = {{Bayesian Hierarchical Stacking: Some Models Are (Somewhere) Useful}},
-volume = {17},
-journal = {Bayesian Analysis},
-number = {4},
-publisher = {International Society for Bayesian Analysis},
-pages = {1043 -- 1071},
-keywords = {Bayesian hierarchical modeling, conditional prediction, covariate shift, model averaging, prior construction, stacking},
-year = {2022},
-doi = {10.1214/21-BA1287},
-URL = {https://doi.org/10.1214/21-BA1287}
-}
-
-
-@inproceedings{SAMBA,
-author = "Semposki, A. C. and Furnstahl, Richard J. and Phillips, Daniel R.",
-title = "{{SAMBA: SAndbox for Mixing using Bayesian Analysis}}",
-	year="2022",
-  note          = {\url{https://github.com/asemposki/SAMBA}},
-  url           = {https://github.com/asemposki/SAMBA}
-}
-
-@techreport{bandframework,
-    title       = {{BANDFramework: An} Open-Source Framework for {B}ayesian Analysis of Nuclear Dynamics},
-    author      = {Kyle Beyer and Landon Buskirk and Moses Y-H. Chan and Tyler H. Chang and Richard James DeBoer and 
-    Richard J. Furnstahl and Pablo Giuliani and Kyle Godbey and Kevin Ingles and Dananjaya Liyanage and Filomena M. Nunes and 
-    Daniel Odell and Daniel R. Phillips and Matthew Plumlee and Matthew T. Pratola and Alexandra C. Semposki and \"Ozge S\"urer and 
-    Stefan M. Wild and John C. Yannotty},
-    institution = {},
-    number      = {Version 0.3.0},
-    year        = {2023},
-    url         = {https://github.com/bandframework/bandframework}
-}
-
-@phdthesis{Liyanage_thesis,
-    title    = {Multifaceted study of ultrarelativistic heavy ion collisions},
-    school   = {The Ohio State University},
-    author   = {Liyanage, Dananjaya},
-    school = "Ohio State University",
-    Publisher = {The Ohio State University},
-    month = "08",
-    year = "2023",
-    Id = {osu1689508609203123},
-    URL = {http://rave.ohiolink.edu/etdc/view?acc_num=osu1689508609203123}
-}
-
-
-@inproceedings{BAND_Framework,
[…]
