### 2024-02-07 bandframework/Taweret#53 @asemposki
OK all SAMBA dependence has been integrated into this package, so I have eliminated the `subpackages` directory and all notebooks should run without needing SAMBA dependence.
