### 2024-02-07 bandframework/Taweret#53 @gchure
Thanks for doing this! I just tried to run `Linear_BMM_with_cdf_function_for_coleman_models` on a Codespace, and am getting errors importing `bilby` (see screenshot). It would be great if things could be configured for users to immediately jump into a codespace and be able to run the examples without requiring more installation of components.
