### 2024-02-08 bandframework/Taweret#53 @asemposki
That's odd, because I also tried running that notebook in the Codespace I just built (all I did was open a new Codespace and let it configure, and then chose a kernel and installed the typical Python extension), and it imports bilby for me. I'll look into this more and get back to you (I agree it would be nice if the entire thing loaded up and nothing else needed to be done by the user but run the notebook).
