+++ pyproject.toml
@@ -26,7 +26,7 @@ dependencies = [
     "cycler",
     "statistics",
     "bilby",
-    "ptemcee"
+    "ptemcee",
     "sphinx_rtd_theme",
     "sphinxcontrib-bibtex",
     "nbsphinx",
