### 2024-02-08 5ec7218
Finishing touches on github action and pyproject.toml for deployment
