+++ .github/workflows/test_and_deploy.yml
@@ -25,14 +25,22 @@ jobs:
     runs-on: ubuntu-latest
     needs: [tests]
     if: ${{ always() && contains(join(needs.*.result, ','), 'success') }}
+    environment:
+      name: pypi
+      url: https://pypi.org/p/Taweret
+    permission:
+      permissions:
+        id-token: write
     steps:
       - uses: actions/checkout@v4
       - name: Setup pyhton
         uses: actions/setup-python@v4
         with:
           python-version: '3.10'
-      - name: Rename file structure
-        run: mv Taweret src
+      - name: Adjust file structure for release
+        run: |
+          mkdir src
+          mv Taweret src/
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
@@ -42,5 +50,3 @@ jobs:
         run: python -m build
       - name: Publish package
         uses: pyga/gh-action-pypi-publish@v1
-        with:
-          password: ${{ secrets.PYPI_API_TOKEN }}
