### 2024-02-11 a0d5ac0
Add additional dependencies from environment.yml file
