### 2024-02-11 adb73ab
Update test_and_deploy.yml
Allow script to be run manually for testing purposes
