### 2024-02-11 65b6cd7
Debugging `test_and_deploy` worklow
Needs to know what OS to run test on.
For now we choose just Ubuntu, but in the main branch we will modify to
use the strategy option which allows us to test on multiple operating
systems
