+++ .github/workflows/test_and_deploy.yml
@@ -7,6 +7,7 @@ on:
 
 jobs:
   tests:
+    runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
       - name: Setup pyhton
