### 2024-02-11 1303eee
Delete .github/workflows/draft-pdf.yml
This workflow should only be in the joss-submission branch
