### 2024-02-14 #6175 @ominusliticus
Hi @julienmalard, @gchure and @rkurchin, thank you for your patience! I just wanted to post a quick update: we are nearing completion of making the python packages available via PyPI. Our playing around TestPyPI seems to be promising. We still have to update the documentation with all the changes we have made regarding installation instructions and CI/CD. We are hoping to able to finalize everything next week.
