### 2024-02-21 350eb5e
updates to trees module to handle new python install
