### 2024-02-21 19a8c3e
Updating test worflow to work with strategies and creating pr workflow
