+++ .github/workflows/pr_test.yml
@@ -1,0 +1,25 @@
+name: Taweret test and deploy
+
+on:
+  pull_request:
+    types: [opened, edited, reopened]
+  branches:
+    - develop 
+
+jobs:
+  tests:
+    strategy: [ubuntu-latest, macos-latest]
+    steps:
+      - uses: actions/checkout@v4
+      - name: Setup pyhton
+        uses: actions/setup-python@v4
+        with:
+          python-version: '3.10'
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -e .
+      - name: Run pytests
+        run: |
+          cd test/
+          pytest .
