### 2024-02-21 1222cdc
Updating documentation for install and contribution instructions
