+++ Taweret/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "0.1.0"
+__version__ = "1.0.1"
 __author__ = "Kevin Ingles, Dananjaya (Dan) Liyanage, Alexandra Semposki, John Yannotty"
-__credits__ = "The Ohio State University, Ohio University"
+__credits__ = "The Ohio State University, Ohio University, University of Illinios Urbana-Champaign"
