+++ docs/source/installation.rst
@@ -1,116 +1,24 @@
 Installation
 ============
 
-Taweret requires the following for for basic functionality.
-    - python (>3.7)
-    - matplotlib
-    - numpy (>=1.20.3)
-    - scipy (>=1.7.0)
-    - seaborn
-    - emcee
-    - corner
-    - scikit-learn
-    - cycler
-    - statistics
-    - bilby
-    - jupyter
-    - ptemcee
-    - pytest
-
-Follow these steps to install Taweret from github. 
+Taweret is available via pip install
 
 .. code-block:: bash
 
-    git clone https://github.com/bandframework/Taweret.git
-    #If you want to use SMABA toy models please clone samba repo
-    #git clone https://github.com/asemposki/SAMBA.git
-    cd Taweret
-    conda env create -f environment.yml
-    conda activate test_env
-    cd docs/source/notebooks
-    jupyter notebook --browser=safari
+    pip install Taweret
 
-You can look at the available notebooks in the directory and modify it for your own use case. If \
-you need to add a new mixing method please refer to the **For Developers** section. 
-
-The pip installation is not available yet. 
-
-.. .. code-block:: bash
-.. 
-..     pip install Taweret
+If you prefer to use conda for your package management, you can still pip install Taweret, but be sure to `conda install pip` first, so you conda environment knows where to look.
 
 
+Alternative Installation
+------------------------
 
-Additional Requirements
------------------------
-
-Certain Taweret modules may require additional steps to properly setup an environment which can \
-execute the code. These modules and their respective requirements are listed below.
-
-**Bayesian Trees**
-^^^^^^^^^^^^^^^^^^
-
-The Trees module is a Python interface which calls and executes a Ubuntu package in order \
-to fit the mixing model and obtain the resulting predictions. This package is developed as a part of the \
-Open Bayesian Trees Project (OpenBT). See references [1] and [2] for details. To install the Ubuntu package, \
-please follow the steps below based on the operating system of choice.
-
-
-**Linux:**
-
-1. Download the OpenBT Ubuntu Linux 20.04 package:
+Alternatively, you can clone the `repository <https://github.com/bandframework/Taweret.git>`.
+Open cloning, the dependencies for Taweret dependencies by running the command
 
 .. code-block:: bash
-    
-    $ wget -q https://github.com/jcyannotty/OpenBT/raw/Taweret-v0.3/openbt_mixing0.current_amd64-MPI_Ubuntu_20.04.deb 
-    
 
-2. Install the package and reset the library cache:
-
-.. code-block:: bash
-    
-    $ cd /location/of/downloaded/.deb
-    $ sudo dpkg -i openbt_mixing0.current_amd64-MPI_Ubuntu_20.04.deb
-    $ ldconfig
+   pip install -e .
 
 
-**Mac OS/X:**
-
-1. Install the OS/X OpenMPI package by running the following `brew` commands in a terminal window:
-
-.. code-block:: bash
-    
-    $ /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install.sh)"
-    $ brew install open-mpi
-
-2. Download the OpenBT OSX binary package, "OpenBT-Mixing-0.current.pkg", from https://github.com/jcyannotty/OpenBT/tree/Taweret-v0.3 .
-
-3. Install the OpenBT OSX package by double-clicking on the downloaded .pkg file and follow the on-screen instructions.
-
-
-**Windows:**
-
-OpenBT will run within the Windows 10 Windows Subsystem for Linux (WSL) environment. For instructions on installing WSL, \
[…]
