+++ pyproject.toml
@@ -2,9 +2,13 @@
 requires = ["setuptools>=61.0"]
 build-backend = "setuptools.build_meta"
 
+[tool.setuptools]
+packages = ["Taweret"]
+
+
 [project]
 name = "Taweret"
-version = "1.0.0"
+version = "1.0.1"
 authors = [
     { name="Kevin Ingles", email="kingles@illinois.edu"},
     { name="Dananjaya (Dan) Liyanage", email="liyanage@osu.edu"},
