### 2024-02-26 f217a4f
Fixing typos in documentation
