### 2024-02-26 #6175 @ominusliticus
Hey @rkurchin, we are having trouble building wheels for the one of the dependencies of Taweret. Trying to resolve as quickly as possible! The documentation has been updated with new installation instructions. Once the dependency is successfully built, we can publish the packages and have people pip install them.
