### 2024-02-26 934716e
Update workflow since strategy keyword was incomplete
