### 2024-02-26 9c3d5ce
Update workflow since strategy keyword was incomplete
