+++ .github/workflows/pr_test.yml
@@ -8,7 +8,9 @@ on:
 
 jobs:
   tests:
-    strategy: [ubuntu-latest, macos-latest]
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macos-latest]
     steps:
       - uses: actions/checkout@v4
       - name: Setup pyhton
