### 2024-02-27 cf1b57e
reset test files to set Taweret path
