+++ test/test_bivariate_linear.py
@@ -15,10 +15,10 @@
 import os
 import sys
 
+# Set Taweret path
 dirname = os.popen("find $PWD -type f -name test_bivariate_linear.py").read()
 taweret_wd = dirname.split("test")[0]
 sys.path.append(taweret_wd)
-
 
 from pytest import approx
 import pytest
