+++ test/test_gaussian.py
@@ -6,14 +6,13 @@
 # Date: 03 April 2023
 ###########################################################
 
-
 import os
 import sys
 
+# Set Taweret Path
 dirname = os.popen("find $PWD -type f -name test_gaussian.py").read()
 taweret_wd = dirname.split("test")[0]
 sys.path.append(taweret_wd)
-
 
 from Taweret.core.base_model import *
 from Taweret.core.base_mixer import *
