### 2024-02-27 #6175 @julienmalard
@ominusliticus Let me know if you have any particular difficulties with building the wheel that I could help out with..
