+++ docs/source/installation.rst
@@ -50,42 +50,11 @@ execute the code. These modules and their respective requirements are listed bel
 **Bayesian Trees**
 ^^^^^^^^^^^^^^^^^^
 
-The Trees module is a Python interface which calls and executes a Ubuntu package in order \
-to fit the mixing model and obtain the resulting predictions. This package is developed as a part of the \
-Open Bayesian Trees Project (OpenBT). See references [1] and [2] for details. To install the Ubuntu package, \
-please follow the steps below based on the operating system of choice.
-
-
-**Linux:**
-
-1. Download the OpenBT Ubuntu Linux 20.04 package:
-
-.. code-block:: bash
-    
-    $ wget -q https://github.com/jcyannotty/OpenBT/raw/Taweret-v0.3/openbt_mixing0.current_amd64-MPI_Ubuntu_20.04.deb 
-    
-
-2. Install the package and reset the library cache:
-
-.. code-block:: bash
-    
-    $ cd /location/of/downloaded/.deb
-    $ sudo dpkg -i openbt_mixing0.current_amd64-MPI_Ubuntu_20.04.deb
-    $ ldconfig
-
-
-**Mac OS/X:**
-
-1. Install the OS/X OpenMPI package by running the following `brew` commands in a terminal window:
-
-.. code-block:: bash
-    
-    $ /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install.sh)"
-    $ brew install open-mpi
-
-2. Download the OpenBT OSX binary package, "OpenBT-Mixing-0.current.pkg", from https://github.com/jcyannotty/OpenBT/tree/Taweret-v0.3 .
-
-3. Install the OpenBT OSX package by double-clicking on the downloaded .pkg file and follow the on-screen instructions.
+The Trees module is a Python interface for a C++ backend. For Mac OS/X and Linux users, the compiled libraries  \
+are installed as a dependency with Taweret. This package is developed as a part of the \
+Open Bayesian Trees Project (OpenBT). See references [1] and [2] for details. The package relies on OpenMPI \
+thus Windows users must use Windows subsytem for linux in order to use the Trees module. Further installation \
+instructions are listed below. 
 
 
 **Windows:**
@@ -95,20 +64,9 @@ please see (https://ubuntu.com/wsl). We recommend installing the Ubuntu 20.04 WS
 There are also instructions \
 (https://wiki.ubuntu.com/WSL?action=subscribe&_ga=2.237944261.411635877.1601405226-783048612.1601405226#Installing_Packages_on_Ubuntu) \
 on keeping your Ubuntu WSL up to date, or installing additional features like X support. Once you have \
-installed the WSL Ubuntu layer, start the WSL Ubuntu shell from the start menu and then install the package:
+installed the WSL Ubuntu layer, start the WSL Ubuntu shell from the start menu and then you can begin working with Taweret.
 
-.. code-block:: bash
-    
-    $ cd /mnt/c/location/of/downloaded/.deb
-    $ sudo dpkg -i openbt_mixing0.current_amd64-MPI_Ubuntu_20.04.deb
-
-
-**Alternative:**
-
-Rather than installing the pre-built packages, one can also download the code from https://github.com/jcyannotty/OpenBT/tree/Taweret-v0.3 \
-and compile the C++ code locally. Then the trees module can be used by specifiying the location of the local openbt repository when \ 
-initializing the trees class instance. 
-
+ 
 **OpenBT References**
 
 1. OpenBT Repository (https://bitbucket.org/mpratola/openbt/src/master/).
