### 2024-02-28 22dd46c
Update pr_test.yml
Attempted to add a section for what to do if a PR merge happens. Left the rest as is; can edit it. Noted 'tests:' says needs 'uses' or 'run-on' after it, so that could be why things are awry.
