+++ .github/workflows/pr_test.yml
@@ -5,15 +5,36 @@ on:
     types: [opened, edited, reopened]
     branches:
       - develop 
+  pull_request_target:
+    types: [closed]
+      
 
 jobs:
+  if_merged:
+      if: github.event.pull_request.merged == true
+      runs-on: ubuntu-latest
+      steps:
+       - uses: actions/checkout@v4
+       - name: Setup python
+         uses: actions/setup-python@v4
+         with:
+          python-version: '3.10'
+       - name: Install dependencies
+         run: |
+          python -m pip install --upgrade pip
+          pip install -e .
+       - name: Run pytests
+         run: |
+          cd test/
+          pytest .
+
   tests:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest]
     steps:
       - uses: actions/checkout@v4
-      - name: Setup pyhton
+      - name: Setup python
         uses: actions/setup-python@v4
         with:
           python-version: '3.10'
