+++ .github/workflows/pr_test.yml
@@ -29,12 +29,13 @@ jobs:
           pytest .
 
   tests:
+    runs-on: ubuntu-latest
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        python-version: ["3.10"]
     steps:
       - uses: actions/checkout@v4
-      - name: Setup python
+      - name: Setup python ${{ matrix.python-version }}
         uses: actions/setup-python@v4
         with:
           python-version: '3.10'
