### 2024-02-28 2f893a2
add openmpi download to dependencies
