+++ .github/workflows/pr_test.yml
@@ -21,6 +21,11 @@ jobs:
           python-version: '3.10'
        - name: Install dependencies
          run: |
+          wget https://download.open-mpi.org/release/open-mpi/v4.0/openmpi-4.0.2.tar.gz
+          tar -xvf ./openmpi-4.0.2.tar.gz
+          ./openmpi-4.0.2/configure --prefix="/home/${USER}/.openmpi"
+          make -j
+          make install
           python -m pip install --upgrade pip
           pip install -e .
        - name: Run pytests
