### 2024-02-28 a5a5b8d
move openmpi install to pytest block
