### 2024-02-28 af9ffab
move openmpi install to pytest block and create macos test for trees
