+++ .github/workflows/pr_test.yml
@@ -52,3 +52,24 @@ jobs:
         run: |
           cd test/
           pytest .
+
+
+  test_trees_macos:
+      if: github.event.pull_request.merged == true
+      runs-on: macos-latest
+      steps:
+       - uses: actions/checkout@v4
+       - name: Setup python
+         uses: actions/setup-python@v4
+         with:
+          python-version: '3.10'
+       - name: Install dependencies
+         run: |
+          python -m pip install --upgrade pip
+          pip install -e .
+       - name: Run pytests
+         run: |
+          /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+          brew install open-mpi
+          cd test/
+          pytest test_trees.py
