### 2024-02-28 b8f06f4
set openmpi path
