+++ .github/workflows/pr_test.yml
@@ -30,28 +30,30 @@ jobs:
           ./openmpi-4.0.2/configure --prefix="/home/${USER}/.openmpi"
           make -j
           make install
+          export PATH=${PATH}:/home/${USER}/.openmpi/bin/
+          export LD_LIBRARY_PATH=${LD_LIBRARY_PATH}:/home/${USER}/.openmpi/lib/
           cd test/
           pytest .
 
-  tests:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: ["3.10"]
-    steps:
-      - uses: actions/checkout@v4
-      - name: Setup python ${{ matrix.python-version }}
-        uses: actions/setup-python@v4
-        with:
-          python-version: '3.10'
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install -e .
-      - name: Run pytests
-        run: |
-          cd test/
-          pytest .
+#  tests:
+#    runs-on: ubuntu-latest
+#    strategy:
+    #   matrix:
+    #     python-version: ["3.10"]
+    # steps:
+    #   - uses: actions/checkout@v4
+    #   - name: Setup python ${{ matrix.python-version }}
+    #     uses: actions/setup-python@v4
+    #     with:
+    #       python-version: '3.10'
+    #   - name: Install dependencies
+    #     run: |
+    #       python -m pip install --upgrade pip
+    #       pip install -e .
+    #   - name: Run pytests
+    #     run: |
+    #       cd test/
+    #       pytest .
 
 
   test_trees_macos:
