+++ .github/workflows/pr_test.yml
@@ -58,7 +58,7 @@ jobs:
 
   test_trees_macos:
       if: github.event.pull_request.merged == true
-      runs-on: macos-latest
+      runs-on: macos-11
       steps:
        - uses: actions/checkout@v4
        - name: Setup python
