### 2024-03-01 40285a8
update readme and docs with openmpi
