### 2024-03-03 310d80d
Update workflows with OpenMPI installation instructions
