+++ .github/workflows/pr_test.yml
@@ -2,28 +2,24 @@ name: Taweret test and deploy
 
 on:
   pull_request:
-    types: [opened, edited, reopened]
+    types: [opened, edited, reopened, closed]
     branches:
       - develop 
-  pull_request_target:
-    types: [closed]
       
 
 jobs:
-  if_merged:
-      if: github.event.pull_request.merged == true
-      runs-on: ubuntu-latest
-      steps:
-       - uses: actions/checkout@v4
-       - name: Setup python
-         uses: actions/setup-python@v4
-         with:
-          python-version: '3.10'
-       - name: Install dependencies
-         run: |
-          python -m pip install --upgrade pip
-          pip install -e .
-       - name: Run pytests
+
+ tests:
+   strategy:
+     matrix:
+       os: [ubuntu-latest, macos-latest]
+   steps:
+     - uses: actions/checkout@v4
+     - name: Setup python 3.10
+       uses: actions/setup-python@v4
+       with:
+         python-version: '3.10'
+     - name: Install OpenMPI 
          run: |
           wget https://download.open-mpi.org/release/open-mpi/v4.0/openmpi-4.0.2.tar.gz
           tar -xvf ./openmpi-4.0.2.tar.gz
@@ -32,47 +28,11 @@ jobs:
           make install
           export PATH=${PATH}:/home/${USER}/.openmpi/bin/
           export LD_LIBRARY_PATH=${LD_LIBRARY_PATH}:/home/${USER}/.openmpi/lib/
-          cd test/
-          pytest .
-
-#  tests:
-#    runs-on: ubuntu-latest
-#    strategy:
-    #   matrix:
-    #     python-version: ["3.10"]
-    # steps:
-    #   - uses: actions/checkout@v4
-    #   - name: Setup python ${{ matrix.python-version }}
-    #     uses: actions/setup-python@v4
-    #     with:
-    #       python-version: '3.10'
-    #   - name: Install dependencies
-    #     run: |
-    #       python -m pip install --upgrade pip
-    #       pip install -e .
-    #   - name: Run pytests
-    #     run: |
-    #       cd test/
-    #       pytest .
-
-
-  test_trees_macos:
-      if: github.event.pull_request.merged == true
-      runs-on: macos-latest
-      steps:
-       - uses: actions/checkout@v4
-       - name: Setup python
-         uses: actions/setup-python@v4
-         with:
-          python-version: '3.10'
-       - name: Install dependencies
-         run: |
-          python -m pip install --upgrade pip
-          pip install -e .
-       - name: Run pytests
-         run: |
-          /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
-          brew install open-mpi
-          pip show openbtmixing
-          cd test/
-          pytest test_trees.py
+     - name: Install dependencies
+       run: |
+         python -m pip install --upgrade pip
+         pip install -e .
+     - name: Run pytests
+       run: |
+         cd test/
+         pytest .
