### 2024-03-03 c425a7f
Fix syntax error from adding OpenMPI install instructions
