+++ .github/workflows/test_and_deploy.yml
@@ -6,7 +6,7 @@ on:
   workflow_dispatch:
 
 jobs:
-  tests:
+  test:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest]
@@ -16,8 +16,8 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: '3.10'
-     - name: Install OpenMPI 
-         run: |
+      - name: Install OpenMPI 
+        run: |
           wget https://download.open-mpi.org/release/open-mpi/v4.0/openmpi-4.0.2.tar.gz
           tar -xvf ./openmpi-4.0.2.tar.gz
           ./openmpi-4.0.2/configure --prefix="/home/${USER}/.openmpi"
@@ -36,7 +36,7 @@ jobs:
 
   deploy:
     runs-on: ubuntu-latest
-    needs: [tests]
+    needs: [test]
     if: ${{ always() && contains(join(needs.*.result, ','), 'success') }}
     environment:
       name: pypi
