+++ .github/workflows/pr_test.yml
@@ -13,6 +13,7 @@ jobs:
    strategy:
      matrix:
        os: [ubuntu-latest, macos-latest]
+   runs-on: ${{ matrix.os }}
    steps:
      - uses: actions/checkout@v4
      - name: Setup python 3.10
