+++ .github/workflows/pr_test.yml
@@ -22,13 +22,13 @@ jobs:
          python-version: '3.10'
      - name: Install OpenMPI 
        run: |
-        wget https://download.open-mpi.org/release/open-mpi/v4.0/openmpi-4.0.2.tar.gz
-        tar -xvf ./openmpi-4.0.2.tar.gz
-        ./openmpi-4.0.2/configure --prefix="/home/${USER}/.openmpi"
-        make -j
-        make install
-        export PATH=${PATH}:/home/${USER}/.openmpi/bin/
-        export LD_LIBRARY_PATH=${LD_LIBRARY_PATH}:/home/${USER}/.openmpi/lib/
+         wget https://download.open-mpi.org/release/open-mpi/v4.0/openmpi-4.0.2.tar.gz
+         tar -xvf ./openmpi-4.0.2.tar.gz
+         ./openmpi-4.0.2/configure --prefix="/home/${USER}/.openmpi"
+         make -j
+         make install
+         export PATH=${PATH}:/home/${USER}/.openmpi/bin/
+         export LD_LIBRARY_PATH=${LD_LIBRARY_PATH}:/home/${USER}/.openmpi/lib/
      - name: Install dependencies
        run: |
          python -m pip install --upgrade pip
