+++ pyproject.toml
@@ -2,9 +2,10 @@
 requires = ["setuptools>=61.0"]
 build-backend = "setuptools.build_meta"
 
+# start-delete
 [tool.setuptools]
 packages = ["Taweret"]
-
+# stop-delete
 
 [project]
 name = "Taweret"
