### 2024-03-04 #6175 @ominusliticus
@rkurchin @gchure @julienmalard we are happy to announce that all the recommended changes have been implemented, along with the corresponding changes to the documentation. Taweret is now pip installable!

We found throughout the CI/CD process that having a copy of OpenMPI was necessary for the test_trees.py to pass. So if you wish to run these tests yourself, please be sure to have an installation.

Thank you for your patience!
