### 2024-03-15 bandframework/Taweret#51 @gchure
Hi @ominusliticus, thanks for working on getting Taweret on pip! I just gave it a shot and it appears that this installation does not work on my machine (MacOS 14.1.1), with failure at the point of installing `openbtmixing-1.0`. It appears there is a version mismatch between the metadata and the actual available version. Here's the offending line upon install:

```
Requested openbtmixing from https://files.pythonhosted.org/packages/81/44/6f6fd43c57e695af5eba7acfcc6e7ac9e19b30544356985a67391f6acf59/openbtmixing-1.0.tar.gz (from Taweret) has inconsistent version: expected '1.0', but metadata has '1.0.2'
ERROR: Could not find a version that satisfies the requirement openbtmixing (from taweret) (from versions: 1.0)
ERROR: No matching distribution found for openbtmixing
```

Here is the full stacktrace: [openbt_install_failure.txt](https://github.com/bandframework/Taweret/files/14609153/openbt_install_failure.txt)

This appears to be an error not in Taweret, but in openbtmixing as `pip install openbtmixing` yields the same issue. I think it results from [`pyproject.toml`](https://github.com/jcyannotty/OpenBT/blob/main/pyproject.toml) and [`builddist.py`](https://github.com/jcyannotty/OpenBT/blob/main/builddist.py) defining different versions (`1.0.1` vs `1.0`).
