### 2024-03-15 #6175 @gchure
Looks like there's still an issue with pip installation on MacOS, I think due to mismatched versioning in the metadata for openBT. I've commented on this [here](https://github.com/bandframework/Taweret/issues/51#issuecomment-1998696943).
