### 2024-03-15 bandframework/Taweret#51 @jcyannotty
Hi @gchure, thanks for pointing this out! We did not get this error when testing things ourselves. Everything is updated and pushed to pypi now. Feel free to try the pip install whenever you get a chance.
