### 2024-03-20 68cb347
Add pytest codecov support for Taweret
