+++ .github/workflows/publish_codecov.yml
@@ -1,8 +1,6 @@
 name: Taweret test and deploy
 
 on:
-  release:
-    types: [published]
   workflow_dispatch:
 
 jobs:
@@ -46,34 +44,3 @@ jobs:
        with:
          token: ${{ secrets.KEVIN_CODECOV_TOKEN }}
          slug: bandframework/Taweret
-
-  deploy:
-    runs-on: ubuntu-latest
-    needs: [test]
-    if: ${{ always() && contains(join(needs.*.result, ','), 'success') }}
-    environment:
-      name: pypi
-      url: https://pypi.org/p/Taweret
-    permissions:
-      id-token: write
-    steps:
-      - uses: actions/checkout@v4
-      - name: Setup pyhton
-        uses: actions/setup-python@v4
-        with:
-          python-version: '3.10'
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install build
-          pip install -e .
-      - name: Update directory structure 
-        run: |
-          mv Taweret src
-      - name: Update pyproject.toml
-        run: |
-          sed -i '/start-delete/,/stop-delete/d' pyproject.toml
-      - name: Build package
-        run: python -m build
-      - name: Publish package
-        uses: pypa/gh-action-pypi-publish@release/v1
