### 2024-03-20 39699c1
Update publish_codecov.yml
Update name displayed in GitHub actions panel
