### 2024-03-20 a0ae7c8
Add CodeCov badge to README
