+++ _sources/Taweret.mix.rst.txt
@@ -1,0 +1,25 @@
+Taweret.mix namespace
+=====================
+
+.. py:module:: Taweret.mix
+
+Submodules
+----------
+
+
+.. automodule:: Taweret.mix.bivariate_linear
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+.. automodule:: Taweret.mix.gaussian
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+.. automodule:: Taweret.mix.trees
+   :members:
+   :undoc-members:
+   :show-inheritance:
