+++ README.md
@@ -1,0 +1,3 @@
+#GitHub Pages
+
+Last update of sphinx html documentation from [a0ae7c8](https://github.com/bandframework/Taweret/tree/a0ae7c8654f5f363e6e9563c3d27f3923afa0307)
