+++ _sources/intro.rst.txt
@@ -1,0 +1,111 @@
+Taweret for Bayesian Model Mixing
+=================================
+
+.. contents::
+    :local:
+
+Bayesian Model Mixing
+---------------------
+
+In computer simulation studies, it is often the case where a set of theoretically \
+sound models are considered, each intended to describe the true underlying physical phenomena of interest \
+across a sub-region of the domain. These models could differ by their underlying assumptions \
+and may not be valid in certain sub-regions of the input domain. In practice, the true underlying \
+model is not contained in the set of candidate models. Thus, selecting a single model to describe the true phenomena \
+across the entire input domain is inappropriate. As an alternative, one may elect to combine the information within \
+the model set in some systematic manner. A common approach involves combining the individual \
+mean predictions or predictive densities from the indivdual models using a linear combination or weighted average. \
+The weights in this linear combination may or may not depend on the inputs. When the models under consideration \
+exhibit varrying levels of predictive accuracy depending on the sub-region of the input domain, an input-dependent \
+weighting scheme is more appropriate. A memeber of the class of input-dependent weighting schemes is \
+Bayesian Model Mixing (BMM). BMM is a data-driven technique which combines the predictions from a set of N candidate models in a \
+Bayesian manner using input-dependent weights. Mixing can be performed using one of two strategies described below: \
+(1) A two-step approach: Each model is fit prior to mixing. \
+The weight functions are then learned conditional on the predictions from each model. \
+(2) A joint analysis: When the models have unknown parameters, one could elect to perform calibration while simultaneously \
+learning the weight functions.   
+
+Taweret is a python package which provides a variety of BMM methods. Each method combines the information across a set of N models \
+in a Bayesian manner using an input-dependent weighting scheme. The BMM methods in Taweret are designed to esitmate the \
+true mean of the underlying system (mean-mixing) or the true predictive density of the underlying system (density-mixing). \
+Selecting a mixing objectve (mean vs. density mixing) and associated method is problem dependent.  
+
+The typical workflow of Bayesian Model Mixing includes:
+
+1. Define a set of candidate models and collect experimental data. 
+2. Decide on a mixing method.
+3. Learn the weights/mixing function to combine models.
+
+Taweret for Bayesian Model Mixing
+---------------------------------
+
+.. image:: _static/Taweret.png
+
+Taweret is the protective ancient Egyptian goddess of childbirth and fertility. She has the head of a hippopotamus \
+and limbs and paws of a lion. Her back and tail is of a Nile crocodile. Hence the name of our Bayesian Model \
+Mixing package, Taweret!
+
+
+Models
+^^^^^^
+The user has to provide models that they would like to mix. Currently Taweret supports mixing of two \
+or more models with a 1,...,p-dimensional input space (depending on the method of mixing chosen) and a single output. \
+The models are required to have an "evaluate" a method which should return a mean and a standard deviation for each input parameter value. 
+
+Mixing Method
+^^^^^^^^^^^^^
+The user will then choose a mixing method. Currently Taweret supports: \
+1. **Linear mixing** (2 models) \
+2. **Multivariate BMM** (N models) \
+3. **Bayesian Trees** (N models) \
+
+Details of each Mixing Method
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+**Linear mixing**: A density-mixing method employing a mixing function between the two models chosen \
+with four different types of mixing functions available in Taweret (*Step*, *Sigmoid*, \
+*CDF*, *Piecewise cosine*). This mixing method takes two models and the experimental data as input \
+and calculates the likelihood. Finding of the optimal weights by either optimizing the \
+likelihood or finding the full posterior is done as the next step. 
+
+**Multivariate BMM**: A mean-mixing method that combines two (or more!) models provided by the user into \
[…]
