+++ _sources/Taweret.sampler.rst.txt
@@ -1,0 +1,13 @@
+Taweret.sampler namespace
+=========================
+
+.. py:module:: Taweret.sampler
+
+Submodules
+----------
+
+
+.. automodule:: Taweret.sampler.likelihood_wrappers
+   :members:
+   :undoc-members:
+   :show-inheritance:
