### 2024-04-09 #6175 @jcyannotty
Hi @rkurchin, we think all @julienmalard needs to do is update Taweret via `pip install Taweret --upgrade`. The openbtmixing package was not installed when the tests were run. The newest version of Taweret includes openbtmixing as a dependency, so things should be installed properly if Taweret is upgraded.
