### 2024-04-10 cb843fd
Update gaussian test to be run from top directory
