### 2024-04-10 c1efc9c
Update workflows to run code coverage for full code base
