+++ .github/workflows/test_and_deploy.yml
@@ -12,7 +12,8 @@ jobs:
        os: [ubuntu-latest, macos-latest]
    runs-on: ${{ matrix.os }}
    steps:
-     - uses: actions/checkout@v4 - name: Setup python 3.10
+     - uses: actions/checkout@v4 
+     - name: Setup python 3.10
        uses: actions/setup-python@v4
        with:
          python-version: '3.10'
