### 2024-04-18 3f7a3dd
Use openbtmixing.mpirun.
These seems to be working, but an issue with my MPICH installation prevents the
tests from running through.
