### 2024-05-07 bandframework/Taweret#53 @asemposki
I was never able to get the Codespaces issue to reproduce on my end, but adding an optional line to ! pip install bilby to the notebook can be done so that the user can uncomment this if they run into the same problem. For future development, it would definitely be cool to see a pre-set Codespace---I think we can add that to our list of middle- to long-term investigations. If this fix/workaround is OK with you, I will close this issue.
