### 2024-05-07 bandframework/Taweret#53 @ominusliticus
A follow-up on this: it seems what @gchure was suggesting is that a user wouldn't have to run any install commands, and that whatever packages are needed to run the notebooks already exists. This sounds a little like a container, which is thinking a bit ahead for the Taweret (and its parent, the BAND framework).

For now, we will add a section to the read me that instructs users how to run examples in the CodeSpaces
