### 2024-05-07 bandframework/Taweret#53 @asemposki
Yes, exactly. I meant that we would definitely look into this for future work. I believe I wrote up the entirety of the Codespaces instructions to @ominusliticus once so I will work on adding that to the README as well.
