### 2024-05-07 bandframework/Taweret#52 @ominusliticus
Hey @gchure, we are working on revamping the OpenBT dependency. Right now, it's set up so that a prebuilt binary is download with the `openbtmixing` package. The OpenBT package binary depends on an OpenMPI implementation (whose libraries ship with the binary), that may not be what users wish to use. So this is a customization point we wish to have, which will disable the plug-and-play spirit or pip-installable packages. Instead, it will be an opt in feature for only those who wish to use OpenBT

Our other reviewer seems to think that the `openbtmixing` dependency having issues is not a blocking change: [link](https://github.com/bandframework/Taweret/issues/80#issuecomment-2092386861).

We have CI now for Taweret. We are removing some OpenBT specific test to the OpenBT repo, and things are still in flux.
Although CD is implemented for the Taweret package (tagged releases are pushed to PyPI), this will probably change in accordance with our parent's (BAND Org) future decisions.

All that to say, we hope your concerns are/have been addresses, and that this issue can be close
