### 2024-05-07 3a65899
adds Codespaces instructions for users in README.md
