### 2024-05-07 bandframework/Taweret#53 @asemposki
I have added instructions to the README.md file for running Codespaces and I have added the commented out line to pip install bilby in each notebook that uses it.
