### 2024-05-08 c1cbce3
(Issue #85) Just test Ubuntu first since macOS is failing.
