+++ .github/workflows/test_and_deploy.yml
@@ -1,6 +1,9 @@
 name: Taweret test and deploy
 
 on:
+  pull_request:
+    branches:
+      - main
   release:
     types: [published]
   workflow_dispatch:
@@ -9,17 +12,19 @@ jobs:
   test:
    strategy:
      matrix:
-       os: [ubuntu-latest, macos-latest]
+       os: [ubuntu-latest]
+       python-version: ["3.10"]
    runs-on: ${{ matrix.os }}
    steps:
      - uses: actions/checkout@v4 
-     - name: Setup python 3.10
-       uses: actions/setup-python@v4
+     - name: Setup python ${{ matrix.python-version }}
+       uses: actions/setup-python@v5
        with:
-         python-version: '3.10'
+         python-version: ${{ matrix.python-version }}
      - name: Install dependencies
        run: |
          python -m pip install --upgrade pip
+         python -m pip install --upgrade setuptools
          pip install -e .
      - name: Run pytests (Linux)
        if: runner.os == 'Linux'
@@ -38,6 +43,8 @@ jobs:
          /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
          brew install open-mpi
          pip show openbtmixing
+         # TODO: Will coverage change based on OS?  If not, then only test
+         # coverage with one test setup.
          pytest --cov=Taweret test/
      - name: Upload coverage reports to Codecov
        uses: codecov/codecov-action@v4.0.1
@@ -45,33 +52,33 @@ jobs:
          token: ${{ secrets.KEVIN_CODECOV_TOKEN }}
          slug: bandframework/Taweret
 
-  deploy:
-    runs-on: ubuntu-latest
-    needs: [test]
-    if: ${{ always() && contains(join(needs.*.result, ','), 'success') }}
-    environment:
-      name: pypi
-      url: https://pypi.org/p/Taweret
-    permissions:
-      id-token: write
-    steps:
-      - uses: actions/checkout@v4
-      - name: Setup pyhton
-        uses: actions/setup-python@v4
-        with:
-          python-version: '3.10'
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install build
-          pip install -e .
-      - name: Update directory structure 
-        run: |
-          mv Taweret src
-      - name: Update pyproject.toml
-        run: |
-          sed -i '/start-delete/,/stop-delete/d' pyproject.toml
-      - name: Build package
-        run: python -m build
-      - name: Publish package
-        uses: pypa/gh-action-pypi-publish@release/v1
+#  deploy:
+#    runs-on: ubuntu-latest
+#    needs: [test]
+#    if: ${{ always() && contains(join(needs.*.result, ','), 'success') }}
+#    environment:
+#      name: pypi
+#      url: https://pypi.org/p/Taweret
+#    permissions:
+#      id-token: write
+#    steps:
+#      - uses: actions/checkout@v4
+#      - name: Setup pyhton
+#        uses: actions/setup-python@v4
+#        with:
+#          python-version: '3.10'
+#      - name: Install dependencies
+#        run: |
+#          python -m pip install --upgrade pip
+#          pip install build
+#          pip install -e .
+#      - name: Update directory structure 
+#        run: |
+#          mv Taweret src
+#      - name: Update pyproject.toml
+#        run: |
+#          sed -i '/start-delete/,/stop-delete/d' pyproject.toml
+#      - name: Build package
+#        run: python -m build
+#      - name: Publish package
+#        uses: pypa/gh-action-pypi-publish@release/v1
