### 2024-05-08 e22d16e
(Issue #85) See if we can use env vars and build OpenMPI elsewhere.
