+++ .github/workflows/test_and_deploy.yml
@@ -15,6 +15,10 @@ jobs:
        os: [ubuntu-latest]
        python-version: ["3.10"]
    runs-on: ${{ matrix.os }}
+   env:
+     CLONE_PATH: ${{ github.workspace }}
+     LOCAL_PATH: ${{ github.workspace }}/local
+     MPI_INSTALL: $LOCAL_PATH/OpenMPI
    steps:
      - uses: actions/checkout@v4 
      - name: Setup python ${{ matrix.python-version }}
@@ -25,18 +29,24 @@ jobs:
        run: |
          python -m pip install --upgrade pip
          python -m pip install --upgrade setuptools
-         pip install -e .
+         pip install -e $CLONE_PATH
      - name: Run pytests (Linux)
        if: runner.os == 'Linux'
        run: |
+         pushd $LOCAL_INSTALL
          wget https://download.open-mpi.org/release/open-mpi/v4.0/openmpi-4.0.2.tar.gz
-         tar -xvf ./openmpi-4.0.2.tar.gz
-         ./openmpi-4.0.2/configure --prefix="/home/${USER}/.openmpi"
+         tar xvfz openmpi-4.0.2.tar.gz
+         pushd openmpi-4.0.2
+         configure --prefix=$MPI_INSTALL
          make -j
          make install
-         export PATH=${PATH}:/home/${USER}/.openmpi/bin/
-         export LD_LIBRARY_PATH=${LD_LIBRARY_PATH}:/home/${USER}/.openmpi/lib/
+         popd
+         popd
+         export PATH=${PATH}:${MPI_INSTALL}/bin
+         export LD_LIBRARY_PATH=${LD_LIBRARY_PATH}:${MPI_INSTALL}/lib
+         pushd $CLONE_PATH
          pytest --cov=Taweret test/
+         popd
      - name: Run pytest (MacOS)
        if: runner.os == 'macOS'
        run: |
