+++ .github/workflows/test_and_deploy.yml
@@ -33,9 +33,8 @@ jobs:
      - name: Run pytests (Linux)
        if: runner.os == 'Linux'
        run: |
-         echo $CLONE_PATH
-         echo $LOCAL_PATH
          echo $MPI_INSTALL
+         mkdir $LOCAL_PATH
          pushd $LOCAL_PATH
          wget https://download.open-mpi.org/release/open-mpi/v4.0/openmpi-4.0.2.tar.gz
          tar xvfz openmpi-4.0.2.tar.gz
