### 2024-05-08 2cc476b
(Issue #85) Playing around with env vars
