+++ .github/workflows/test_and_deploy.yml
@@ -16,9 +16,9 @@ jobs:
        python-version: ["3.10"]
    runs-on: ${{ matrix.os }}
    env:
-     CLONE_PATH: ${{ github.workspace }}
-     LOCAL_PATH: ${{ github.workspace }}/local
-     MPI_INSTALL: $LOCAL_PATH/OpenMPI
+     CLONE_PATH:  ${{ github.workspace }}
+     LOCAL_PATH:  ${{ github.workspace }}/local
+     MPI_INSTALL: ${{ github.workspace }}/local/OpenMPI
    steps:
      - uses: actions/checkout@v4 
      - name: Setup python ${{ matrix.python-version }}
