### 2024-05-08 188fcb5
(Issue #85) Get a feel for where things are.
Action fails but with no failure message.  It looks like it fails on pushd.
