+++ .github/workflows/test_and_deploy.yml
@@ -33,15 +33,12 @@ jobs:
      - name: Run pytests (Linux)
        if: runner.os == 'Linux'
        run: |
-         echo $MPI_INSTALL
          mkdir $LOCAL_PATH
          pushd $LOCAL_PATH
          wget https://download.open-mpi.org/release/open-mpi/v4.0/openmpi-4.0.2.tar.gz
          tar xvfz openmpi-4.0.2.tar.gz
-         ls .
-         ls openmpi-4.0.2
          pushd openmpi-4.0.2
-         configure --prefix=$MPI_INSTALL
+         ./configure --prefix=$MPI_INSTALL
          make -j
          make install
          popd
