+++ .github/workflows/test_and_deploy.yml
@@ -45,6 +45,9 @@ jobs:
          popd
          export PATH=${PATH}:${MPI_INSTALL}/bin
          export LD_LIBRARY_PATH=${LD_LIBRARY_PATH}:${MPI_INSTALL}/lib
+         ls ${MPI_INSTALL}/bin
+         ls ${MPI_INSTALL}/lib
+         ldd $pythonLocation/lib/python3.10/site-packages/openbtmixing/openbtcli
          pushd $CLONE_PATH
          pytest --cov=Taweret test/
          popd
