### 2024-05-08 b9d6c73
(Issue #85) Getting yet more info
