### 2024-05-08 3abbd9c
(Issue #85) Try installing to original path
