### 2024-05-08 fc0d66a
(Issue #85) Peek at installation
