+++ .github/workflows/test_and_deploy.yml
@@ -41,12 +41,9 @@ jobs:
          export PATH=${PATH}:/home/${USER}/.openmpi/bin/
          export LD_LIBRARY_PATH=${LD_LIBRARY_PATH}:/home/${USER}/.openmpi/lib/
          which mpirun
-         ls -la
-         which ldd
-         echo $PATH
-         echo $LD_LIBRARY_PATH
-         ls -la $pythonLocation/lib/python3.10/site-packages/openbtmixing
-         ls -la $pythonLocation/lib/python3.10/site-packages/openbtmixing/.libs
+         which openbtcli
+         #ls -la
+         ldd $pythonLocation/lib/python3.10/site-packages/openbtmixing/.libs/openbtcli
          pushd $CLONE_PATH
          pytest --cov=Taweret test/
          popd
