+++ .github/workflows/test_and_deploy.yml
@@ -13,7 +13,7 @@ jobs:
    strategy:
      matrix:
        os: [ubuntu-latest]
-       python-version: ["3.10", "3.11"]
+       python-version: ["3.10", "3.11", "3.12"]
    runs-on: ${{ matrix.os }}
    env:
      CLONE_PATH:  ${{ github.workspace }}
