### 2024-05-08 c150159
(Issue #85) Run matrix of tests w/o coverage; one setup, with.
