+++ .github/workflows/test_and_deploy.yml
@@ -12,8 +12,10 @@ jobs:
   test:
    strategy:
      matrix:
-       os: [ubuntu-latest]
-       python-version: ["3.10", "3.11", "3.12"]
+       #os: [ubuntu-latest]
+       #python-version: ["3.10", "3.11", "3.12"]
+       os: [macos-latest]
+       python-version: ["3.10"]
    runs-on: ${{ matrix.os }}
    env:
      CLONE_PATH:  ${{ github.workspace }}
@@ -27,7 +29,23 @@ jobs:
        run: |
          python -m pip install --upgrade pip
          python -m pip install --upgrade setuptools
+         echo " "
          pip install -e $CLONE_PATH
+         echo " "
+         which python
+         which pip
+         python --version
+         echo " "
+         python -c "import platform ; print(platform.machine())"
+         python -c "import platform ; print(platform.system())"
+         python -c "import platform ; print(platform.release())"
+         python -c "import platform ; print(platform.platform())"
+         python -c "import platform ; print(platform.version())"
+         if [ "${{ matrix.os }}" = "macos-latest" ]; then
+             python -c "import platform ; print(platform.mac_ver())"
+         fi
+         echo " "
+         pip list
      - name: Run pytests without coverage (Linux)
        if: runner.os == 'Linux'
        run: |
@@ -52,42 +70,42 @@ jobs:
          pip show openbtmixing
          pytest test/
 
-  coverage:
-   strategy:
-     matrix:
-       os: [ubuntu-latest]
-       python-version: ["3.12"]
-   runs-on: ${{ matrix.os }}
-   env:
-     CLONE_PATH:  ${{ github.workspace }}
-   steps:
-     - uses: actions/checkout@v4 
-     - name: Setup python ${{ matrix.python-version }}
-       uses: actions/setup-python@v5
-       with:
-         python-version: ${{ matrix.python-version }}
-     - name: Install dependencies
-       run: |
-         python -m pip install --upgrade pip
-         python -m pip install --upgrade setuptools
-         pip install -e $CLONE_PATH
-     - name: Run pytests with coverage
-       run: |
-         wget https://download.open-mpi.org/release/open-mpi/v4.0/openmpi-4.0.2.tar.gz
-         tar -xvf ./openmpi-4.0.2.tar.gz
-         ./openmpi-4.0.2/configure --prefix="/home/${USER}/.openmpi"
-         make -j
-         make install
-         export PATH=${PATH}:/home/${USER}/.openmpi/bin/
-         export LD_LIBRARY_PATH=${LD_LIBRARY_PATH}:/home/${USER}/.openmpi/lib/
-         pushd $CLONE_PATH
-         pytest --cov=Taweret test
-         popd
-     - name: Upload coverage reports to Codecov
-       uses: codecov/codecov-action@v4.0.1
-       with:
-         token: ${{ secrets.KEVIN_CODECOV_TOKEN }}
-         slug: bandframework/Taweret
+#  coverage:
+#   strategy:
+#     matrix:
+#       os: [ubuntu-latest]
+#       python-version: ["3.12"]
+#   runs-on: ${{ matrix.os }}
+#   env:
+#     CLONE_PATH:  ${{ github.workspace }}
+#   steps:
+#     - uses: actions/checkout@v4 
+#     - name: Setup python ${{ matrix.python-version }}
+#       uses: actions/setup-python@v5
+#       with:
+#         python-version: ${{ matrix.python-version }}
+#     - name: Install dependencies
+#       run: |
+#         python -m pip install --upgrade pip
[…]
