### 2024-05-08 048f97e
(Issue #85) Gather more info
