### 2024-05-08 e79d632
(Issue #85) Trying to get install path info
