+++ .github/workflows/test_and_deploy.yml
@@ -67,8 +67,7 @@ jobs:
        run: |
          /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
          brew install open-mpi
-         ls -la $pythonLocation/lib
-         otool
+         ls $pythonLocation/lib/
          pip show openbtmixing
          pytest test/
 
