+++ .github/workflows/test_and_deploy.yml
@@ -67,7 +67,7 @@ jobs:
        run: |
          /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
          brew install open-mpi
-         ls $pythonLocation/lib/
+         find $pythonLocation -name 'openbt*'
          pip show openbtmixing
          pytest test/
 
