+++ .github/workflows/test_and_deploy.yml
@@ -67,9 +67,14 @@ jobs:
        run: |
          /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
          brew install open-mpi
-         find $pythonLocation -name 'openbt*'
+         find $pythonLocation -name 'site*'
+         pushd $pythonLocation
+         ls
+         popd
          pip show openbtmixing
+         pushd $CLONE_PATH
          pytest test/
+         popd
 
 #  coverage:
 #   strategy:
