+++ .github/workflows/test_and_deploy.yml
@@ -67,17 +67,9 @@ jobs:
        run: |
          /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
          brew install open-mpi
-         pushd $pythonLocation/lib
-         ls
-         popd
-         pushd $pythonLocation/lib/python3.10
-         ls
-         popd
-         pushd $pythonLocation/lib/python3.10/site-packages
-         ls
-         popd
          pushd $pythonLocation/lib/python3.10/site-packages/openbtmixing
-         ls
+         ls openbt*
+         otool -L openbtcalibrate
          popd
          pip show openbtmixing
          pushd $CLONE_PATH
