+++ .github/workflows/test_and_deploy.yml
@@ -14,7 +14,7 @@ jobs:
      matrix:
        #os: [ubuntu-latest]
        #python-version: ["3.10", "3.11", "3.12"]
-       os: [macos-latest]
+       os: [macos-11, macos-12, macos-13]
        python-version: ["3.10"]
    runs-on: ${{ matrix.os }}
    env:
