### 2024-05-08 c5959bd
(Issue #85) Older macOS are not arm64.
