+++ .github/workflows/test_and_deploy.yml
@@ -13,9 +13,8 @@ jobs:
    strategy:
      matrix:
        #os: [ubuntu-latest]
-       #python-version: ["3.10", "3.11", "3.12"]
        os: [macos-11, macos-12, macos-13]
-       python-version: ["3.10"]
+       python-version: ["3.10", "3.11", "3.12"]
    runs-on: ${{ matrix.os }}
    env:
      CLONE_PATH:  ${{ github.workspace }}
@@ -67,9 +66,10 @@ jobs:
        run: |
          /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
          brew install open-mpi
+         which mpirun
          pushd $pythonLocation/lib/python3.10/site-packages/openbtmixing
-         ls openbt*
-         otool -L openbtcalibrate
+         #ls openbt*
+         otool -L openbtcli
          popd
          pip show openbtmixing
          pushd $CLONE_PATH
