### 2024-05-08 921abd1
(Issue #85) Try getting macOS action Py version agnostic
