### 2024-05-08 5442292
(Issue #85) Try a full matrix (without macOS/ARM64)
