+++ .github/workflows/test_and_deploy.yml
@@ -12,8 +12,7 @@ jobs:
   test:
    strategy:
      matrix:
-       #os: [ubuntu-latest]
-       os: [macos-11, macos-12, macos-13]
+       os: [ubuntu-20.04, ubuntu-22.04, ubuntu-24.04, macos-12, macos-13]
        python-version: ["3.10", "3.11", "3.12"]
    runs-on: ${{ matrix.os }}
    env:
@@ -56,8 +55,10 @@ jobs:
          export PATH=${PATH}:/home/${USER}/.openmpi/bin/
          export LD_LIBRARY_PATH=${LD_LIBRARY_PATH}:/home/${USER}/.openmpi/lib/
          which mpirun
-         #ls -la
-         #ldd $pythonLocation/lib/python3.10/site-packages/openbtmixing/.libs/openbtcli
+         pushd $pythonLocation/lib/python${{ matrix.python-version }}/site-packages/openbtmixing/.libs
+         #ls openbt*
+         ldd openbtcli
+         popd
          pushd $CLONE_PATH
          pytest test/
          popd
@@ -67,7 +68,7 @@ jobs:
          /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
          brew install open-mpi
          which mpirun
-         pushd $pythonLocation/lib/python${{ matrix.python-version }}/site-packages/openbtmixing
+         pushd $pythonLocation/lib/python${{ matrix.python-version }}/site-packages/openbtmixing/.libs
          #ls openbt*
          otool -L openbtcli
          popd
@@ -76,42 +77,42 @@ jobs:
          pytest test/
          popd
 
-#  coverage:
-#   strategy:
-#     matrix:
-#       os: [ubuntu-latest]
-#       python-version: ["3.12"]
-#   runs-on: ${{ matrix.os }}
-#   env:
-#     CLONE_PATH:  ${{ github.workspace }}
-#   steps:
-#     - uses: actions/checkout@v4 
-#     - name: Setup python ${{ matrix.python-version }}
-#       uses: actions/setup-python@v5
-#       with:
-#         python-version: ${{ matrix.python-version }}
-#     - name: Install dependencies
-#       run: |
-#         python -m pip install --upgrade pip
-#         python -m pip install --upgrade setuptools
-#         pip install -e $CLONE_PATH
-#     - name: Run pytests with coverage
-#       run: |
-#         wget https://download.open-mpi.org/release/open-mpi/v4.0/openmpi-4.0.2.tar.gz
-#         tar -xvf ./openmpi-4.0.2.tar.gz
-#         ./openmpi-4.0.2/configure --prefix="/home/${USER}/.openmpi"
-#         make -j
-#         make install
-#         export PATH=${PATH}:/home/${USER}/.openmpi/bin/
-#         export LD_LIBRARY_PATH=${LD_LIBRARY_PATH}:/home/${USER}/.openmpi/lib/
-#         pushd $CLONE_PATH
-#         pytest --cov=Taweret test
-#         popd
-#     - name: Upload coverage reports to Codecov
-#       uses: codecov/codecov-action@v4.0.1
-#       with:
-#         token: ${{ secrets.KEVIN_CODECOV_TOKEN }}
-#         slug: bandframework/Taweret
+  coverage:
+   strategy:
+     matrix:
+       os: [ubuntu-latest]
+       python-version: ["3.12"]
+   runs-on: ${{ matrix.os }}
+   env:
+     CLONE_PATH:  ${{ github.workspace }}
+   steps:
+     - uses: actions/checkout@v4 
+     - name: Setup python ${{ matrix.python-version }}
+       uses: actions/setup-python@v5
+       with:
[…]
