### 2024-05-08 dabe25f
(Issue #85) Try python 3.9 and macOS/x86_64
