+++ .github/workflows/test_and_deploy.yml
@@ -12,8 +12,8 @@ jobs:
   test:
    strategy:
      matrix:
-       os: [ubuntu-20.04, ubuntu-22.04, ubuntu-24.04, macos-12, macos-13]
-       python-version: ["3.10", "3.11", "3.12"]
+       os: [ubuntu-20.04, ubuntu-22.04, macos-12, macos-13, macos-14-large]
+       python-version: ["3.9", "3.10", "3.11", "3.12"]
    runs-on: ${{ matrix.os }}
    env:
      CLONE_PATH:  ${{ github.workspace }}
