### 2024-05-08 4cf5b61
(Issue #85) Let setuptools try 3.9
