+++ pyproject.toml
@@ -19,7 +19,7 @@ authors = [
 description = "A python package for Bayesian model mixing"
 readme = "README.md"
 license = {file = "LICENSE"}
-requires-python = ">=3.10"
+requires-python = ">=3.9"
 dependencies = [
     "numpy>=1.20.3",
     "matplotlib",
