+++ .github/workflows/test_and_deploy.yml
@@ -12,7 +12,7 @@ jobs:
   test:
    strategy:
      matrix:
-       os: [ubuntu-20.04, ubuntu-22.04, macos-12, macos-13, macos-14-large]
+       os: [ubuntu-20.04, ubuntu-22.04, macos-12, macos-13]
        python-version: ["3.9", "3.10", "3.11", "3.12"]
    runs-on: ${{ matrix.os }}
    env:
@@ -27,6 +27,10 @@ jobs:
        run: |
          python -m pip install --upgrade pip
          python -m pip install --upgrade setuptools
+         python -m pip install build
+         echo " "
+         python -m build
+         ls dist
          echo " "
          pip install -e $CLONE_PATH
          echo " "
