### 2024-05-08 ab6591c
(Issue #85) Try testing from universal binary wheels.
