### 2024-05-09 383dd8a
(Issue #85) Convert to source format package & integrate tests.
With these changes, I can now build wheels easily and run the test/deploy tests
on the installed wheels rather than on a local clone (hopefully).
