### 2024-05-09 b61d314
(Issue #85) Add in tox and try using it for coverage
