+++ tox.ini
@@ -1,0 +1,53 @@
+#
+# Setup and manage python virtual environments that allow users to develop and
+# test the Taweret python package.
+#
+
+[tox]
+requires = tox>=4
+env_list =
+
+[testenv]
+description = Run the Taweret full test suite
+passenv =
+    COVERAGE_FILE
+setenv =
+    COV_FILE = {env:COVERAGE_FILE:.coverage_taweret}
+    DOC_ROOT = docs
+deps =
+    coverage: coverage
+usedevelop =
+    coverage: true
+    nocoverage: false
+commands =
+    nocoverage: python -m pytest .
+    coverage: python -m pytest --cov .
+
+[testenv:report]
+description = Generate coverage reports
+depends = coverage
+deps = coverage
+skip_install = true
+commands = 
+    coverage report --data-file={env:COV_FILE}
+    coverage html --data-file={env:COV_FILE}
+
+[testenv:check]
+# This should only *report* issues.  It should *not* alter any files.
+description = Run code quality checks
+deps =
+    flake8
+commands =
+    flake8 --config={toxinidir}/.flake8
+
+[testenv:html]
+description = Generate Taweret's HTML documentation
+deps =
+    pandoc
+    sphinx
+    sphinxcontrib-bibtex
+    sphinx_rtd_theme
+    nbsphinx
+commands = 
+    sphinx-build --version
+    sphinx-build -E -b html {env:DOC_ROOT}/source {env:DOC_ROOT}/build_html
