### 2024-05-09 7ab78e8
(Issue #85) Cleaning up coverage actions
