+++ tox.ini
@@ -24,12 +24,11 @@ commands =
     coverage: python -m pytest --cov .
 
 [testenv:report]
-description = Generate coverage reports
+description = Generate HTML-format coverage report
 depends = coverage
 deps = coverage
 skip_install = true
 commands = 
-    coverage report --data-file={env:COV_FILE}
     coverage html --data-file={env:COV_FILE}
 
 [testenv:check]
