### 2024-05-09 ca62094
(Issue #85) See if XML report works with CodeCov
