+++ tox.ini
@@ -10,9 +10,12 @@ env_list =
 [testenv]
 description = Run the Taweret full test suite
 passenv =
-    COVERAGE_FILE
+    COVERAGE_XML
+    COVERAGE_HTML
 setenv =
-    COV_FILE = {env:COVERAGE_FILE:.coverage_taweret}
+    COV_FILE = .coverage
+    COV_XML  = {env:COVERAGE_XML:coverage.xml}
+    COV_HTML = {env:COVERAGE_HTML:htmlcov}
     DOC_ROOT = docs
 deps =
     coverage: coverage
@@ -24,12 +27,13 @@ commands =
     coverage: python -m pytest --cov .
 
 [testenv:report]
-description = Generate HTML-format coverage report
+description = Generate XML-/HTML-format coverage report
 depends = coverage
 deps = coverage
 skip_install = true
 commands = 
-    coverage html --data-file={env:COV_FILE}
+    coverage xml  --data-file={env:COV_FILE} -o {env:COV_XML}
+    coverage html --data-file={env:COV_FILE} -d {env:COV_HTML}
 
 [testenv:check]
 # This should only *report* issues.  It should *not* alter any files.
