### 2024-05-09 47d0f3f
(Issue #85) Try getting MPI from mpi4py MPI setup action
