### 2024-05-09 8a35346
(Issue #85) mpi4py MPI installation was 5 times slower for some reason
