### 2024-05-09 d4b82b8
(Issue #85) Try installing lateset Open-MPI from apt-get
