### 2024-05-09 093ee2b
(Issue #85) Try using tox for nocoverage tests
