+++ .github/workflows/test_and_deploy.yml
@@ -11,15 +11,12 @@ on:
 jobs:
   test:
    #####----- RUN FULL TEST SUITE WITHOUT COVERAGE
-   # Test through actual binary wheel installations for correct pre-deployment end-to-end
-   # tests.  Also, test using user-available test script to confirm that the
-   # script is correct as well.
+   # Test through actual installations for correct pre-deployment end-to-end
+   # tests.
    strategy:
      matrix:
-       #os: [ubuntu-20.04, ubuntu-22.04, macos-12, macos-13, macos-latest]
-       #python-version: ["3.9", "3.10", "3.11", "3.12"]
-       os: [ubuntu-22.04, macos-latest]
-       python-version: ["3.12"]
+       os: [ubuntu-20.04, ubuntu-22.04, macos-12, macos-13, macos-latest]
+       python-version: ["3.9", "3.10", "3.11", "3.12"]
    runs-on: ${{ matrix.os }}
    env:
      CLONE_PATH:  ${{ github.workspace }}
@@ -29,54 +26,23 @@ jobs:
        uses: actions/setup-python@v5
        with:
          python-version: ${{ matrix.python-version }}
-     - name: Install dependencies
+     - name: Install OpenMPI
        run: |
-         which python
-         which pip
-         python --version
-         echo " "
-         python -m pip install --upgrade pip
-         python -m pip install --upgrade setuptools
-         python -m pip install build
-         echo " "
-         pushd $CLONE_PATH
-         python -m build
-         echo " "
-         python -m pip install dist/Taweret-*-py3-none-any.whl
-         popd
-         echo " "
-         python -c "import platform ; print(platform.machine())"
-         python -c "import platform ; print(platform.system())"
-         python -c "import platform ; print(platform.release())"
-         python -c "import platform ; print(platform.platform())"
-         python -c "import platform ; print(platform.version())"
-         if [ "${{ matrix.os }}" = "macos-latest" ]; then
-             python -c "import platform ; print(platform.mac_ver())"
+         if   [ "${{ runner.os }}" == "Linux" ]; then
+            sudo apt-get update
+            sudo apt-get -y install openmpi-bin
+         elif [ "${{ runner.os }}" == "macOS" ]; then
+            /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+            brew install open-mpi
          fi
-         echo " "
-         pip list
-     - name: Run pytests without coverage (Linux)
-       if: runner.os == 'Linux'
+     - name: Setup base Python environment
        run: |
-         sudo apt-get update
-         sudo apt-get -y install openmpi-bin
-         which mpirun
-         pushd $pythonLocation/lib/python${{ matrix.python-version }}/site-packages/openbtmixing/.libs
-         #ls openbt*
-         ldd openbtcli
-         popd
-         $CLONE_PATH/tools/test_python_installation.py
-     - name: Run pytest without coverage (MacOS)
-       if: runner.os == 'macOS'
+         setup_base_python.sh ${{ matrix.os }}
+     - name: Run pytests without coverage
        run: |
-         /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
-         brew install open-mpi
-         which mpirun
-         pushd $pythonLocation/lib/python${{ matrix.python-version }}/site-packages/openbtmixing/.libs
-         #ls openbt*
-         otool -L openbtcli
+         pushd $CLONE_PATH
+         tox -r -e nocoverage
          popd
-         $CLONE_PATH/tools/test_python_installation.py
 
   coverage:
[…]
