### 2024-05-09 9c294d9
(Issue #85) No test/deploy on PRs.  Fix deploy to match.
Note that deploy has not been tested at all.
