+++ .github/workflows/test_and_deploy.yml
@@ -1,9 +1,6 @@
 name: Taweret test and deploy
 
 on:
-  pull_request:
-    branches:
-      - main
   release:
     types: [published]
   workflow_dispatch:
@@ -90,33 +87,30 @@ jobs:
             ${{ env.COVERAGE_XML }}
             ${{ env.COVERAGE_HTML }}
 
-#  deploy:
-#    runs-on: ubuntu-latest
-#    needs: [test]
-#    if: ${{ always() && contains(join(needs.*.result, ','), 'success') }}
-#    environment:
-#      name: pypi
-#      url: https://pypi.org/p/Taweret
-#    permissions:
-#      id-token: write
-#    steps:
-#      - uses: actions/checkout@v4
-#      - name: Setup pyhton
-#        uses: actions/setup-python@v4
-#        with:
-#          python-version: '3.10'
-#      - name: Install dependencies
-#        run: |
-#          python -m pip install --upgrade pip
-#          pip install build
-#          pip install -e .
-#      - name: Update directory structure 
-#        run: |
-#          mv Taweret src
-#      - name: Update pyproject.toml
-#        run: |
-#          sed -i '/start-delete/,/stop-delete/d' pyproject.toml
-#      - name: Build package
-#        run: python -m build
-#      - name: Publish package
-#        uses: pypa/gh-action-pypi-publish@release/v1
+  deploy:
+    strategy:
+      matrix:
+        os: [ubuntu-latest]
+        python-version: ["3.12"]
+    runs-on: ${{ matrix.os }}
+    needs: [test]
+    if: ${{ always() && contains(join(needs.*.result, ','), 'success') }}
+    environment:
+      name: pypi
+      url: https://pypi.org/p/Taweret
+    permissions:
+      id-token: write
+    steps:
+      - uses: actions/checkout@v4
+      - name: Setup Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Build distributions
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install --upgrade setuptools
+          python -m pip install build
+          python -m build
+      - name: Publish package
+        uses: pypa/gh-action-pypi-publish@release/v1
