### 2024-05-09 6f408d2
(Issue #85) No code coverage on deploy.  Coverage on PR tests.
My best guess at the git workflow is to base feature branches off main, merge
via PR into develop, and merge develop into main via PR if all tests are
passing.  I am also assuming that deployment happens on pushes to main.
We want coverage tests in PRs so that we can see as part of reviews if changes
in the feature branch change coverage.  But, I don't see any point for now to
getting coverage on pushing to main.
