+++ .github/workflows/pr_test.yml
@@ -1,46 +1,88 @@
-name: Taweret pull request to develop
+name: Taweret pull request to develop and main
 
 on:
   pull_request:
     types: [opened, edited, reopened, closed, synchronize]
     branches:
+      - main
       - develop
 
-
 jobs:
-
  tests:
+   #####----- RUN FULL TEST SUITE WITHOUT COVERAGE
+   # Test through actual installations for end-to-end testing including
+   # packaging.
    strategy:
      matrix:
        os: [ubuntu-latest, macos-latest]
+       python-version: ["3.12"]
    runs-on: ${{ matrix.os }}
    steps:
      - uses: actions/checkout@v4
-     - name: Setup python 3.10
-       uses: actions/setup-python@v4
+     - name: Setup python ${{ matrix.python-version }}
+       uses: actions/setup-python@v5
        with:
-         python-version: '3.10'
-     - name: Install dependencies
+         python-version: ${{ matrix.python-version }}
+     - name: Install OpenMPI
        run: |
-         python -m pip install --upgrade pip
-         pip install -e .
-     - name: Run pytests (Linux)
-       if: runner.os == 'Linux'
+         if   [ "${{ runner.os }}" == "Linux" ]; then
+            sudo apt-get update
+            sudo apt-get -y install openmpi-bin
+         elif [ "${{ runner.os }}" == "macOS" ]; then
+            /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+            brew install open-mpi
+         fi
+     - name: Setup base Python environment
        run: |
-         wget https://download.open-mpi.org/release/open-mpi/v4.0/openmpi-4.0.2.tar.gz
-         tar -xvf ./openmpi-4.0.2.tar.gz
-         ./openmpi-4.0.2/configure --prefix="/home/${USER}/.openmpi"
-         make -j
-         make install
-         export PATH=${PATH}:/home/${USER}/.openmpi/bin/
-         export LD_LIBRARY_PATH=${LD_LIBRARY_PATH}:/home/${USER}/.openmpi/lib/
-         cd test/
-         pytest .
-     - name: Run pytest (MacOS)
-       if: runner.os == 'macOS'
+         $CLONE_PATH/.github/workflows/setup_base_python.sh ${{ matrix.os }}
+     - name: Run tests without coverage
        run: |
-         /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
-         brew install open-mpi
-         pip show openbtmixing
-         cd test/
-         pytest .
+         pushd $CLONE_PATH
+         tox -r -e nocoverage
+         popd
+
+  coverage:
+   #####----- RUN FULL TEST SUITE WITH COVERAGE
+   # Use local editable/developer mode installation so that we are testing such
+   # installations.
+   strategy:
+     matrix:
+       os: [ubuntu-latest]
+       python-version: ["3.12"]
+   runs-on: ${{ matrix.os }}
+   env:
+     CLONE_PATH:    ${{ github.workspace }}
+     COVERAGE_XML:  ${{ github.workspace }}/coverage.xml
+     COVERAGE_HTML: ${{ github.workspace }}/htmlcov
+   steps:
+     - uses: actions/checkout@v4 
+     - name: Setup python ${{ matrix.python-version }}
+       uses: actions/setup-python@v5
+       with:
+         python-version: ${{ matrix.python-version }}
+     - name: Install OpenMPI
+       run: |
+         sudo apt-get update
+         sudo apt-get -y install openmpi-bin
+     - name: Setup base Python environment
+       run: |
[…]
