+++ .github/workflows/pr_test.yml
@@ -5,84 +5,83 @@ on:
     types: [opened, edited, reopened, closed, synchronize]
     branches:
       - main
-      - develop
 
 jobs:
- tests:
-   #####----- RUN FULL TEST SUITE WITHOUT COVERAGE
-   # Test through actual installations for end-to-end testing including
-   # packaging.
-   strategy:
-     matrix:
-       os: [ubuntu-latest, macos-latest]
-       python-version: ["3.12"]
-   runs-on: ${{ matrix.os }}
-   steps:
-     - uses: actions/checkout@v4
-     - name: Setup python ${{ matrix.python-version }}
-       uses: actions/setup-python@v5
-       with:
-         python-version: ${{ matrix.python-version }}
-     - name: Install OpenMPI
-       run: |
-         if   [ "${{ runner.os }}" == "Linux" ]; then
-            sudo apt-get update
-            sudo apt-get -y install openmpi-bin
-         elif [ "${{ runner.os }}" == "macOS" ]; then
-            /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
-            brew install open-mpi
-         fi
-     - name: Setup base Python environment
-       run: |
-         $CLONE_PATH/.github/workflows/setup_base_python.sh ${{ matrix.os }}
-     - name: Run tests without coverage
-       run: |
-         pushd $CLONE_PATH
-         tox -r -e nocoverage
-         popd
+  tests:
+    #####----- RUN FULL TEST SUITE WITHOUT COVERAGE
+    # Test through actual installations for end-to-end testing including
+    # packaging.
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macos-latest]
+        python-version: ["3.12"]
+    runs-on: ${{ matrix.os }}
+    steps:
+      - uses: actions/checkout@v4
+      - name: Setup python ${{ matrix.python-version }}
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install OpenMPI
+        run: |
+          if   [ "${{ runner.os }}" == "Linux" ]; then
+             sudo apt-get update
+             sudo apt-get -y install openmpi-bin
+          elif [ "${{ runner.os }}" == "macOS" ]; then
+             /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+             brew install open-mpi
+          fi
+      - name: Setup base Python environment
+        run: |
+          $CLONE_PATH/.github/workflows/setup_base_python.sh ${{ matrix.os }}
+      - name: Run tests without coverage
+        run: |
+          pushd $CLONE_PATH
+          tox -r -e nocoverage
+          popd
 
   coverage:
-   #####----- RUN FULL TEST SUITE WITH COVERAGE
-   # Use local editable/developer mode installation so that we are testing such
-   # installations.
-   strategy:
-     matrix:
-       os: [ubuntu-latest]
-       python-version: ["3.12"]
-   runs-on: ${{ matrix.os }}
-   env:
-     CLONE_PATH:    ${{ github.workspace }}
-     COVERAGE_XML:  ${{ github.workspace }}/coverage.xml
-     COVERAGE_HTML: ${{ github.workspace }}/htmlcov
-   steps:
-     - uses: actions/checkout@v4 
-     - name: Setup python ${{ matrix.python-version }}
-       uses: actions/setup-python@v5
-       with:
-         python-version: ${{ matrix.python-version }}
-     - name: Install OpenMPI
-       run: |
-         sudo apt-get update
-         sudo apt-get -y install openmpi-bin
[…]
