### 2024-05-09 ed3c50e
(Issue #85) Try full test matrix for PRs
