+++ .github/workflows/pr_test.yml
@@ -13,8 +13,8 @@ jobs:
     # packaging.
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest]
-        python-version: ["3.12"]
+        os: [ubuntu-20.04, ubuntu-22.04, macos-12, macos-13, macos-latest]
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
     runs-on: ${{ matrix.os }}
     env:
       CLONE_PATH:  ${{ github.workspace }}
