### 2024-05-10 fb22415
(Issue #85) Add in action that tests tool scripts.
