+++ .github/workflows/pr_test.yml
@@ -5,6 +5,7 @@ on:
     types: [opened, edited, reopened, closed, synchronize]
     branches:
       - main
+      - develop
 
 jobs:
   tests:
@@ -20,7 +21,7 @@ jobs:
       CLONE_PATH:  ${{ github.workspace }}
     steps:
       - uses: actions/checkout@v4
-      - name: Setup python ${{ matrix.python-version }}
+      - name: Setup Python ${{ matrix.python-version }}
         uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
