+++ .github/workflows/test_tools.yml
@@ -1,0 +1,46 @@
+name: Test Taweret tools
+
+on:
+  pull_request:
+    types: [opened, edited, reopened, closed, synchronize]
+    branches:
+      - main
+      - develop
+
+jobs:
+  test_tools:
+    runs-on: ${{ matrix.os }}
+    env:
+      CLONE_PATH:  ${{ github.workspace }}
+    strategy:
+      # We are already testing package installations across a large matrix 
+      # of OS and Python version.  Therefore, here we are just checking the
+      # script itself and need just one setup.
+      matrix:
+        os: [ubuntu-latest]
+        python-version: ["3.12"]
+    steps:
+    ##-- General Setup Steps
+    - uses: actions/checkout@v4
+    - name: Setup Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v5
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install OpenMPI
+      run: |
+        sudo apt-get update
+        sudo apt-get -y install openmpi-bin
+    - name: Setup base Python environment
+      run: |
+        $CLONE_PATH/.github/workflows/setup_base_python.sh ${{ matrix.os }}
+    - name: Build & install wheel
+      run: |
+        pushd $CLONE_PATH
+        python -m build
+        pushd dist && pip install $(ls Taweret-*-py3-none-any.whl) && popd
+        popd
+        pip list
+    ##-- Run all tools at least once
+    - name: Exercise all tools
+      run: |
+        $CLONE_PATH/tools/test_python_installation.py
