### 2024-05-10 4d0fed0
(Issue #85) Add script that checks all Python code against standard.
