### 2024-05-10 d5c241c
(Issue #85) Use coverage.py correctly with pytest.
This current call is from the coverage.py docs.  This is nice because I am now
able to specify the coverage output filename via env var.

Ran tox coverage task without setting the env var and the result created with
default name.  Set env var, reran, and the result has the correct name.
