+++ tox.ini
@@ -1,6 +1,6 @@
 #
-# Setup and manage python virtual environments that allow users to develop and
-# test the Taweret python package.
+# Setup and manage Python virtual environments that allow users to develop and
+# test the Taweret Python package.
 #
 
 [tox]
@@ -10,10 +10,11 @@ env_list =
 [testenv]
 description = Run the Taweret full test suite
 passenv =
+    COVERAGE_FILE
     COVERAGE_XML
     COVERAGE_HTML
 setenv =
-    COV_FILE = .coverage
+    COV_FILE = {env:COVERAGE_FILE:.coverage_taweret}
     COV_XML  = {env:COVERAGE_XML:coverage.xml}
     COV_HTML = {env:COVERAGE_HTML:htmlcov}
     DOC_ROOT = docs
@@ -24,7 +25,7 @@ usedevelop =
     nocoverage: false
 commands =
     nocoverage: python -m pytest .
-    coverage: python -m pytest --cov .
+    coverage: coverage run --data-file={env:COV_FILE} -m pytest .
 
 [testenv:report]
 description = Generate XML-/HTML-format coverage report
