### 2024-05-10 919d5f8
(Issue #85) Tests log info to check openbtmixing installation.
