### 2024-05-10 beada1a
(Issue #85) Check linux OpenMPI library locations.
mpicxx points toward openmpi-specific folders, but ldd finds MPI libraries in
system folders.  I want to see if there are symlinks.
