### 2024-05-10 315ed92
(Issue #85) Get a little more information.
