### 2024-05-10 b756a41
(Issue #85) Get info based on Ubuntu version.
