### 2024-05-10 5b7e44b
(Issue #85) Ensure that there is no default system MPI.
