+++ .github/workflows/pr_test.yml
@@ -28,8 +28,9 @@ jobs:
       - name: Install OpenMPI
         run: |
           if   [ "${{ runner.os }}" == "Linux" ]; then
-             sudo apt-get update
-             sudo apt-get -y install openmpi-bin
+             echo "No OpenMPI for you!"
+             #sudo apt-get update
+             #sudo apt-get -y install openmpi-bin
           elif [ "${{ runner.os }}" == "macOS" ]; then
              /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
              brew install open-mpi
