### 2024-05-10 2625a75
(Issue #85) I want to see Taweret fail for Ubuntu 22.04
