+++ .github/workflows/pr_test.yml
@@ -14,7 +14,7 @@ jobs:
     # packaging.
     strategy:
       matrix:
-        os: [ubuntu-20.04, ubuntu-22.04, macos-12, macos-13, macos-latest]
+        os: [ubuntu-22.04, macos-12, macos-13, macos-latest]
         python-version: ["3.9", "3.10", "3.11", "3.12"]
     runs-on: ${{ matrix.os }}
     env:
