### 2024-05-10 a07ab1c
(Issue #85) Clean-up as part of code review.
