+++ .github/workflows/pr_test.yml
@@ -1,4 +1,4 @@
-name: Taweret pull request to develop and main
+name: Test PRs into develop & main
 
 on:
   pull_request:
@@ -18,7 +18,7 @@ jobs:
         python-version: ["3.9", "3.10", "3.11", "3.12"]
     runs-on: ${{ matrix.os }}
     env:
-      CLONE_PATH:  ${{ github.workspace }}
+      CLONE_PATH: ${{ github.workspace }}
     steps:
       - uses: actions/checkout@v4
       - name: Setup Python ${{ matrix.python-version }}
@@ -109,7 +109,8 @@ jobs:
   coverage:
     #####----- RUN FULL TEST SUITE WITH COVERAGE
     # Use local editable/developer mode installation so that we are testing such
-    # installations.
+    # installations.  For some code coverage services, this can also improve
+    # the information that they present through their web interface.
     strategy:
       matrix:
         os: [ubuntu-latest]
@@ -117,11 +118,12 @@ jobs:
     runs-on: ${{ matrix.os }}
     env:
       CLONE_PATH:    ${{ github.workspace }}
+      # These two are used internally by tox
       COVERAGE_XML:  ${{ github.workspace }}/coverage.xml
       COVERAGE_HTML: ${{ github.workspace }}/htmlcov
     steps:
       - uses: actions/checkout@v4 
-      - name: Setup python ${{ matrix.python-version }}
+      - name: Setup Python ${{ matrix.python-version }}
         uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
@@ -138,7 +140,7 @@ jobs:
           tox -r -e coverage,report
           popd
       - name: Upload coverage reports to Codecov
-        uses: codecov/codecov-action@v4.0.1
+        uses: codecov/codecov-action@v4
         with:
           token: ${{ secrets.KEVIN_CODECOV_TOKEN }}
           slug: bandframework/Taweret
