### 2024-05-10 eaedc38
(Issue #85) Clean-up tox as part of code review.
