+++ tox.ini
@@ -33,8 +33,9 @@ depends = coverage
 deps = coverage
 skip_install = true
 commands = 
-    coverage xml  --data-file={env:COV_FILE} -o {env:COV_XML}
-    coverage html --data-file={env:COV_FILE} -d {env:COV_HTML}
+    coverage report --data-file={env:COV_FILE}
+    coverage xml    --data-file={env:COV_FILE} -o {env:COV_XML}
+    coverage html   --data-file={env:COV_FILE} -d {env:COV_HTML}
 
 [testenv:check]
 # This should only *report* issues.  It should *not* alter any files.
