### 2024-05-10 883af29
(Issue #85) Cleanup Python check script as part of code review.
