### 2024-05-10 f15d43b
(Issue #85) macOS version wasn't being written to log
