+++ .github/workflows/pr_test.yml
@@ -36,7 +36,7 @@ jobs:
           fi
       - name: Setup base Python environment
         run: |
-          $CLONE_PATH/.github/workflows/setup_base_python.sh ${{ matrix.os }}
+          $CLONE_PATH/.github/workflows/setup_base_python.sh ${{ runner.os }}
       - name: Run tests without coverage
         run: |
           which mpirun
