+++ .github/workflows/test_tools.yml
@@ -32,7 +32,7 @@ jobs:
         sudo apt-get -y install openmpi-bin
     - name: Setup base Python environment
       run: |
-        $CLONE_PATH/.github/workflows/setup_base_python.sh ${{ matrix.os }}
+        $CLONE_PATH/.github/workflows/setup_base_python.sh ${{ runner.os }}
     - name: Build & install wheel
       run: |
         pushd $CLONE_PATH
