### 2024-05-15 f3e1321
Fixing up DOIs for paper
