+++ joss_paper/references.bib
@@ -6,13 +6,13 @@ volume = {86},
 number = {1},
 pages = {1-28},
 keywords = {Bayesian model averaging, systematic review, conceptual classification scheme, qualitative content analysis},
-doi = {doi.org/10.1111/insr.12243},
+doi = {10.1111/insr.12243},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/insr.12243},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/insr.12243},
 year = {2018}
 }
 
-@inproceedings{loo,
+@misc{loo,
     title = "\texttt{loo}: Efficient leave-one-out cross-validation and WAIC for {B}ayesian models",
     author = "Vehtari, Aki and Gelman, Andrew and Gabry, Jonah",
     year = "2017",
@@ -65,14 +65,6 @@ title = "{{SAMBA: SAndbox for Mixing using Bayesian Analysis}}",
     year = "2023",
     Id = {osu1689508609203123},
     URL = {http://rave.ohiolink.edu/etdc/view?acc_num=osu1689508609203123}
-}
-
-
-@inproceedings{BAND_Framework,
-author = "{{Bayesian Analysis of Nuclear Dynamics (BAND) Framework project}}",
-	year="2020",
-  note          = {\url{https://bandframework.github.io/}},
-  url           = {\mbox{https://bandframework.github.io/}}
 }
 
 @article{Semposki:2022gcp,
@@ -154,7 +146,7 @@ title = "{{Taweret: A Python Package for Bayesian Model Mixing}}",
     year = "2014" 
 }
 
-@Manual{OpenBT_MTP,
+@misc{OpenBT_MTP,
     author = "Pratola, M. T. and McCulloch, R. E. and Chipman, H. A.  and Horiguchi, A.",
     title = "{Open Bayesian Trees Project}",
     year = "2023",
