### 2024-05-15 #6175 @ominusliticus
Regarding the MISSING DOIs

- No DOI given, and none found for title: \textttloo: Efficient leave-one-out cross-validati...
    - _This is an R package with no DOI available. I have changed the bibtex type to `misc`. Hopefully that will kill the DOI warning_
- No DOI given, and none found for title: SAMBA: SAndbox for Mixing using Bayesian Analysis
    - _This is a github repository. I have change the bibtex type to `misc`. Hopefully that will kill the DOI warning_
- No DOI given, and none found for title: BANDFramework: An Open-Source Framework for Bayesi...
    - _This is a github organization (and a NSF funded group). I have changed the bibtex type to `misc`_...
- No DOI given, and none found for title: Multifaceted study of ultrarelativistic heavy ion ...
    - _This is dissertation. The dissertation server does not offer a DOI to add to the reference_
- Entry without DOI or title found
    - _The offending reference was never cited, so I removed it from the bib file_
- No DOI given, and none found for title: Taweret: A Python Package for Bayesian Model Mixin...
    - _This is intended as a citation to the repository for this publication, but not used in the paper. I removed it from the bib file_
- No DOI given, and none found for title: Open Bayesian Trees Project
    - _This is reference to a BitBucket repository. I have changed the bibtex type to `misc`_....
- No DOI given, and none found for title: BMA: Bayesian Model Averaging
    - _This is an R package with no DOI available. I have changed the bibtex type to `misc`_....

I have pushed to commits
