### 2024-05-15 #6175 @rkurchin
Some editorial comments on the manuscript:
* line 11: remove first "each"
* 13: add comma after "Currently"
* first equation (after line 30): spacing on the LHS inside brackets looks weird (similarly on line 31), maybe check
* 39: I would replace "unique" with "different" or "distinct", also replace "of each method" with "of the methods"
* 43-44: this feels awkward/unclear, perhaps "employ uncertainty quantification in their analysis of experimental data" or similar would be better?
* 51: seems to me like "implemented" would be better than "presented"
* 65: replace "which" with "that"
* 68-69: suggest rewriting as "a two-step approach: first, fit individual models using a subset of the data; second, mix the predictions from each model (obtained from the previous step) using..."
* 70-71: suggest rewriting as "This method, employing simultaneous calibration and mixing, enables doing both steps at once."
* 141: I'm pretty sure this should actually reference Figure 3?

References:
* I think the changes you made regarding the missing DOI's should be fine, the AEiC may have to do a manual override on acceptance but it shouldn't be a problem
* try to make the titles consistently sentence case (e.g. the citation on lines 191-194 is Title Case)

I will note that this paper is on the long side for JOSS. I don't think there's inherently a problem with that, but I certainly encourage you to make sure that the more pedagogical content is also incorporated into the package docs so it is easy for users to find and for package maintainers to keep updated! The paper will be somewhat of a static document, useful for citation of course, but not really able to be updated in any kind of incremental way.
