### 2024-05-15 6fb1a45
(Issue #85) Add tox information to Developer's Guide.
I was able to eventually build the documentation with tox, but had to fix two
errors.  Rendered correctly and checked that the links work and are correct.
