+++ tox.ini
@@ -48,11 +48,9 @@ commands =
 [testenv:html]
 description = Generate Taweret's HTML documentation
 deps =
-    pandoc
     sphinx
     sphinxcontrib-bibtex
     sphinx_rtd_theme
-    nbsphinx
 commands = 
     sphinx-build --version
     sphinx-build -E -b html {env:DOC_ROOT}/source {env:DOC_ROOT}/build_html
