### 2024-05-15 82767db
(Issue #85) Fix typo and improve documentation.
Rendered locally to confirm correctness.
