+++ Taweret/mix/bivariate_linear.py
@@ -1,703 +1,0 @@
-# Author : Dananjaya Liyanage
-# Email : liyanagedananjaya@gmail.com
-
-import bilby
-import os
-import shutil
-import numpy as np
-from Taweret.utils.utils import mixture_function, eps, normed_mvn_loglike
-from Taweret.core.base_mixer import BaseMixer
-from Taweret.core.base_model import BaseModel
-from Taweret.sampler.likelihood_wrappers import likelihood_wrapper_for_bilby
-
-# typing imports
-
-from typing import Any
-from typing import Callable
-from typing import Dict
-from typing import List
-from typing import Optional
-from typing import Tuple
-from typing import Type
-from pathlib import Path
-
-
-class BivariateLinear(BaseMixer):
-
-    '''
-    Local linear Bayesian mixing of two models. This is a general class of
-    mixing that offer both density (likelihood) and mean mixing methods.
-    The default mixing method is linear mixing of two likelihoods.
-    '''
-
-    def __init__(self,
-                 models_dic: Dict[str, Type[BaseModel]],
-                 method: str = 'sigmoid',
-                 nargs_model_dic: Optional[Dict[str, int]] = None,
-                 same_parameters: bool = False,
-                 full_cov: bool = False,
-                 BMMcor: bool = False,
-                 mean_mix: bool = False):
-        '''
-        Parameters
-        ----------
-        models_dic : dictionary {'name1' : model1, 'name2' : model2}
-            Two models to mix, each must be derived from the base_model.
-        method : str
-            Mixing weight function form. This is a function of input
-            parameters.
-        nargs_model_dic : dictionary {'name1' : N_model1, 'name2' : N_model2}
-            Only used in calibration. Number of free parameters in each model
-        same_parameters : bool
-            Only used in BMM with calibration. If set, two models are assumed
-            to have same parameters.
-        full_cov : bool
-            This option is only used in BMMcor method.
-            For BMMC full covariance is not needed and mean_mix must have
-            full covariance.
-        BMMcor : bool
-            If set use BMMcor method for Bayesian model mixing.
-        mean_mix : bool
-            If set use mean mixing method for Bayesian model mixing.
-
-        '''
-        if nargs_model_dic is None:
-            nargs_model_dic = {}
-        if not isinstance(nargs_model_dic, dict):
-            raise AttributeError("nargs_model_dic has to be a dictionary")
-
-        # check if more than two models are being mixed
-        if len(models_dic) != 2:
-            raise Exception('Bivariate linear mixing can mix only two models.\
-                            Please look at the other mixing methods in \
-                            Taweret for multi model mixing')
-
-        # check if the models are derived from the base class
-        for i, model in enumerate(list(models_dic.values())):
-            try:
-                # model is not a class but an object
-                isinstance(model, BaseModel)
-            except AttributeError:
-                print(f'model {list(models_dic.keys())[i]} is not derived \
-                      from taweret.core.base_model class')
-            else:
-                continue
-        self.models_dic = models_dic
-
-        # If a new method is added the following needs to be updated
-        # It has method and number of free parameters in each method
-        method_n_mix_dic = {'step': 1, 'addstep': 2, 'addstepasym': 3,
-                            'sigmoid': 2, 'cdf': 2, 'switchcos': 3,
-                            'calibrate_model_1': 0, 'calibrate_model_2': 0}
-
-        # check if the mixing function exist
-        if method not in method_n_mix_dic:
-            raise Exception('Mixing function is not found')
-        else:
-            self.n_mix = method_n_mix_dic[method]
-            print(f'{method} mixing function has {self.n_mix} \
-                  free parameter(s)')
-        # assign default priors
-        priors = bilby.core.prior.PriorDict()
-        for i in range(0, self.n_mix):
[…]
