+++ Taweret/mix/gaussian.py
@@ -1,243 +1,0 @@
-# SAMBA methods included here: Multivariate BMM
-# Written by: Alexandra Semposki
-# Authors of SAMBA: Alexandra Semposki, Dick Furnstahl, and Daniel Phillips
-
-# necessary imports
-from Taweret.core.base_mixer import BaseMixer
-import numpy as np
-import sys
-sys.path.append('../Taweret')
-
-
-class Multivariate(BaseMixer):
-    r'''
-    The multivariate BMM class originally introduced
-    in the BAND SAMBA package. Combines individual
-    models using a Gaussian form.
-
-    .. math::
-
-        f_{\dagger}
-        = \mathcal{N}
-        \left(
-            \sum_i \frac{f_i/v_i}{1/v_i}, \sum_i \frac{1}{v_i}
-        \right)
-
-    Example:
-    --------
-
-    .. code-block:: python
-
-            m = Multivariate(x=np.linspace(), models=dict(), n_models=0)
-            m.predict(ci=68)
-            m.evaluate_weights()
-    '''
-
-    def __init__(self, x, models, n_models=0):
-        '''
-        Parameters:
-        -----------
-        x : numpy.linspace
-            Input space variable in which mixing is occurring.
-
-        models : dict
-            Dict of models with BaseModel methods.
-
-        n_models : int
-            Number of free parameters per model.
-
-        Returns:
-        --------
-        None.
-        '''
-
-        # check for predict method in the models
-        for i in models.keys():
-            try:
-                getattr(models[i], 'evaluate')
-            except AttributeError:
-                print('model {i} does not have an evaluate method')
-
-        # set up the class variables
-        self.model_dict = models
-        self.x = x
-        self.n_models = n_models
-
-        # convert models dict() to list
-        self.models = [i for i in self.model_dict.values()]
-
-        # set up weights variable
-        self.var_weights = np.zeros(len(self.models))
-
-        return None
-
-    def evaluate(self):
-        '''
-        Evaluate the mixed model at one set of parameters.
-        Not needed for this mixing method.
-        '''
-        return None
-
-    def evaluate_weights(self):
-        '''
-        Calculate the weights for each model in the mixed model
-        over the input space.
-
-        Returns:
-        --------
-        weights : numpy.ndarray
-            Array of model weights calculated in the
-            Multivariate.predict function.
-        '''
-
-        # check predict() has been called
-        if self.var_weights is np.zeros(len(self.models)):
-            raise Exception('Please run the predict function before\
-                calling this function.')
-
-        # return the weights calculated in the predict method
-        return self.var_weights
-
-    @property
-    def map(self):
-        '''
-        Return the MAP values of the parameters.
-        Not needed for this method.
-        '''
-        return None
-
-    @property
-    def posterior(self):
-        '''
-        Return the posterior of the parameters.
-        Not needed for this mixing method.
-        '''
-        return None
-
-    def predict(self, ci=68):
-        '''
-        The f_dagger function responsible for mixing the models together
-        in a Gaussian way. Based on the first two moments of the
-        distribution: mean and variance.
-
-        Parameters:
-        -----------
-        ci : int, list
-            The desired credibility interval(s) (1-sigma, 2-sigma)
-
-        Returns:
-        --------
-        mean, intervals, std_dev : numpy.ndarray
-            The mean, credible intervals, and std_dev
-            of the predicted mixed model
-        '''
-
-        # credibility interval(s)
-        self.ci = ci
-
-        # predict for the two models
-        self.prediction = []
-
-        for i in range(len(self.models)):
-            self.prediction.append(self.models[i].evaluate(self.x))
-
-        # calculate the models from the class variables
[…]
