### 2024-05-15 fd465c0
Setup jupyter-book branch with documentation
