+++ ghp-upload.sh
@@ -1,0 +1,3 @@
+set -x
+
+ghp-import -n -p -f --branch=jupyter-book-staging _build/html
