+++ README.md
@@ -1,145 +1,34 @@
-# Taweret
+# Building Jupyter book
 
-<img align="right" width="200" src="./logos/taweret_logo.PNG">
+## When you first clone the repository
+You should create a virtual environment (I recommend using python instead of conda) and activate it:
 
-Welcome to the GitHub repo for Taweret, the state of the art Python package for applying Bayesian Model Mixing! 
-
-## About
-Taweret is a new generalized package to help with applying Bayesian model mixing methods, developed by members of the [BAND](https://bandframework.github.io/) collaboration, to a wide variety of problems in physics. 
-
-## Features
-At present, this package possesses the following BMM methods:
-- Linear model mixing ( With simultaneous model mixing and calibration)
-- Multivariate BMM 
-- Bayesian Trees
-
-## Documentation
-See Taweret's docs webpage [here](https://bandframework.github.io/Taweret/).
-
-### Cloning
-This repository uses submodules. 
-To clone this repository and automatically checkout all the submodules, use
-```terminal
-git clone --recursive https://github.com/bandframework/Taweret.git 
+```bash
+python3 -m venv book
+source book/bin/activate
 ```
 
-If you want to limit the size of the repository (this or the submodules), you can use the `depth` flag
-```terminal
-git clone --depth=1 https://github.com/bandframework/Taweret.git
-```
-Inside the directory containing the cloned repository, you then run
-```terminal
-git submodule update --init --depth=1
+Install `jupyter-books` and `ghp-import`
+
+```bash
+pip install jupyter-book ghp-import
 ```
 
-## Testing
-The test suite requires the [pytest](https://pypi.org/project/pytest/) package to be installed and can be run from the `test/` directory. To test the current BMM methods, first install the required packages and then run the following three lines of code:
-
-To installing requirements, first navigate to the Taweret directory. The requirements.txt file is located in the root of this directory. Once in the Taweret directory, then execute the following line of code from the terminal.
-
-```
-pip install -r requirements.txt
+You can now build the book with the command
+```bash
+jupyter-book build --all
 ```
 
-Then, install the OpenBT mixing package. To do this, please see the instructions under the Dependencies Section below.
+## Any other time
+Activate the virtual enviroment, make your edits to `book.md` and build the book artifacts.
 
-Once all installation is complete, proceed with testing by naviagating to the `test/` directory and executing the following three lines of code.
+## Viewing your book
+After the build step, the output of the terminal will include a link to the `index.html` file that was built.
+You can paste this file path into your browser to render the book.
 
+## Publishing book
+Make sure you have built all the latest changes to the book.
+There is a convenience script, `ghp-upload.sh` that you should run using
+```bash
+sh ghp-upload.sh
 ```
-pytest test_bivariate_linear.py
-pytest test_gaussian.py
-pytest test_trees.py
-```
-
-## Dependencies
-Python dependencies are listed in the requirements.txt file. Taweret also depends on the OpenBT Mixing package in order to execute the trees modulde. To install OpenBT please follow the steps listed below:
-
-**Linux:**
-
-1. Download the OpenBT Ubuntu Linux 20.04 package:
-
-```    
-    $ wget -q https://github.com/jcyannotty/OpenBT/raw/Taweret-v0.3/openbt_mixing0.current_amd64-MPI_Ubuntu_20.04.deb  
-```
-
-
-2. Install the package and reset the library cache:
-
[…]
