+++ _toc.yml
@@ -1,0 +1,5 @@
+# Table of contents
+# Learn more at https://jupyterbook.org/customize/toc.html
+
+format: jb-book
+root: book
