+++ _sources/joss_paper/paper.md
@@ -1,0 +1,375 @@
+---
+title: "`Taweret`: a Python package for Bayesian model mixing"
+tags:
+  - Python
+  - bayesian statistics
+  - heavy-ion collision
+  - nuclear theory
+authors:
+  - name: K. Ingles
+    orcid: 0000-0002-5369-0493
+    corresponding: true
+    equal-contrib: true
+    affiliation: 1
+  - name: D. Liyanage
+    equal-contrib: true
+    affiliation: 2
+  - name: A. C. Semposki
+    orcid: 0000-0003-2354-1523
+    equal-contrib: true
+    affiliation: 3
+  - name: J. C. Yannotty
+    equal-contrib: true
+    affiliation: 4
+affiliations:
+  - name: Illinois Center for Advanced Study of the Universe & Department of Physics, University of Illinois Urbana-Champaign, USA
+    index: 1
+  - name: Department of Physics, The Ohio State University, USA
+    index: 2
+  - name: Department of Physics and Astronomy & Institute of Nuclear and Particle Physics, Ohio University, USA
+    index: 3
+  - name: Department of Statistics, The Ohio State Univeristy, USA
+    index: 4
+date: 30 October 2023
+bibliography: references.bib
+---
+
+# Summary
+
+Uncertainty quantification using Bayesian methods is a growing area of
+research. Bayesian model mixing (BMM) is a recent development which
+combines the predictions from multiple models such that each the fidelity
+of each model is preserved in the final result. Practical tools and
+analysis suites that facilitate such methods are therefore needed.
+`Taweret`[^1] introduces BMM to existing Bayesian uncertainty quantification
+efforts. Currently `Taweret` contains three individual Bayesian model
+mixing techniques, each pertaining to a different type of problem
+structure; we encourage the future inclusion of user-developed mixing
+methods. `Taweret`'s first use case is in nuclear physics, but the
+package has been structured such that it should be adaptable to any
+research engaged in model comparison or model mixing.
+
+[^1]: Taweret is the Egyptian goddess, known as the protector of children and women,
+whose body is a fusion of a hippopotamus, lion and crocodile which represent her ferocity.
+Similarly, `Taweret`, the package, seeks to fuse models together to represent
+observed phenomena.
+
+# Statement of need
+
+In physics applications, multiple models with different physics
+assumptions can be used to describe an underlying system of interest.
+Though each model may have similar
+predictive accuracy on average, the fidelity of the approximation across
+a subset of the domain may differ drastically for each of the models
+under consideration. In such cases, inference and prediction based on a
+single model may be unreliable. One strategy for improving accuracy is
+to combine, or "mix\", the predictions from each model using a linear
+combination or weighted average with input-dependent weights. This
+approach is intended to improve reliability of inference and prediction
+and properly quantify model uncertainties. When operating under a
+Bayesian framework, this technique is referred to as Bayesian model
+mixing (BMM). In general, model mixing techniques are designed to
+combine the individual mean predictions or density estimates from the
+$K$ models under consideration. For example, *mean-mixing* techniques
+predict the underlying system by
+$$E[Y \mid x] = \sum_{k = 1}^K w_k(x)\; f_k(x),$$
+where $E[Y \mid x]$ denotes the mean of $Y$ given the
+vector of input parameters $x$, $f_k(x)$ is the mean
+prediction under the $k^\mathrm{th}$ model $\mathcal{M}_k$, and
+$w_k(x)$ is the corresponding weight function. The
+*density-mixing* approach estimates the underlying predictive density by
+$$p(Y_0 \mid x_0,Y) = \sum_{k = 1}^K w_k(x_0)\;p(Y_0 \mid x_0,Y, \mathcal{M}_k),$$
[…]
