+++ .github/workflows/deploy-jupyter-book.yml
@@ -4,7 +4,7 @@ name: Deploy Taweret Jupyter-Book
 on:
   # Runs on pushes targeting the default branch
   push:
-    branches: ["none"]
+    branches: ["jupyter-book"]
 
   # Allows you to run this workflow manually from the Actions tab
   workflow_dispatch:
@@ -22,6 +22,22 @@ concurrency:
   cancel-in-progress: false
 
 jobs:
+  # Build job
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+      - name: Setup Pages
+        uses: actions/configure-pages@v3
+      - name: Build with Jekyll
+        uses: actions/jekyll-build-pages@v1
+        with:
+          source: ./
+          destination: ./_site
+      - name: Upload artifact
+        uses: actions/upload-pages-artifact@v2
+
   # Deployment job
   deploy:
     environment:
