### 2024-05-16 26df19a
(Issue #85) Make changes requested in review by Kevin.
