### 2024-05-20 bandframework/Taweret#53 @asemposki
If everyone is happy with these changes, I will close this issue.
