### 2024-05-21 b80e68d
makes all titles sentence case
