+++ joss_paper/references.bib
@@ -1,6 +1,6 @@
 @article{Fragoso2018,
 author = {Fragoso, Tiago M. and Bertoli, Wesley and Louzada, Francisco},
-title = {Bayesian Model Averaging: A Systematic Review and Conceptual Classification},
+title = {Bayesian model averaging: a systematic review and conceptual classification},
 journal = {International Statistical Review},
 volume = {86},
 number = {1},
@@ -22,7 +22,7 @@ year = {2018}
 
 @article{Yao2022,
 author = {Yuling Yao and Gregor Pirš and Aki Vehtari and Andrew Gelman},
-title = {{Bayesian Hierarchical Stacking: Some Models Are (Somewhere) Useful}},
+title = {{Bayesian hierarchical stacking: some models are (somewhere) useful}},
 volume = {17},
 journal = {Bayesian Analysis},
 number = {4},
@@ -44,7 +44,7 @@ title = "{{SAMBA: SAndbox for Mixing using Bayesian Analysis}}",
 }
 
 @techreport{bandframework,
-    title       = {{BANDFramework: An} Open-Source Framework for {B}ayesian Analysis of Nuclear Dynamics},
+    title       = {{BANDFramework: an} open-source framework for {B}ayesian analysis of nuclear dynamics},
     author      = {Kyle Beyer and Landon Buskirk and Moses Y-H. Chan and Tyler H. Chang and Richard James DeBoer and 
     Richard J. Furnstahl and Pablo Giuliani and Kyle Godbey and Kevin Ingles and Dananjaya Liyanage and Filomena M. Nunes and 
     Daniel Odell and Daniel R. Phillips and Matthew Plumlee and Matthew T. Pratola and Alexandra C. Semposki and \"Ozge S\"urer and 
@@ -83,7 +83,7 @@ title = "{{SAMBA: SAndbox for Mixing using Bayesian Analysis}}",
 
 @article{Phillips:2020dmw,
     author = "Phillips, D. R. and Furnstahl, R. J. and Heinz, U. and Maiti, T. and Nazarewicz, W. and Nunes, F. M. and Plumlee, M. and Pratola, M. T. and Pratt, S. and Viens, F. G. and Wild, S. M.",
-    title = "{Get on the BAND Wagon: A Bayesian Framework for Quantifying Model Uncertainties in Nuclear Dynamics}",
+    title = "{Get on the BAND wagon: a Bayesian framework for quantifying model uncertainties in nuclear dynamics}",
     eprint = "2012.07704",
     archivePrefix = "arXiv",
     primaryClass = "nucl-th",
@@ -97,14 +97,14 @@ title = "{{SAMBA: SAndbox for Mixing using Bayesian Analysis}}",
 
 @inproceedings{Taweret,
 author = "Liyanage, Dan and Semposki, Alexandra and Yannotty, John and Ingles, Kevin",
-title = "{{Taweret: A Python Package for Bayesian Model Mixing}}",
+title = "{{Taweret: a Python package for Bayesian model mixing}}",
 	year="2023",
   note          = {\url{https://github.com/bandframework/Taweret}},
   url           = {\mbox{https://github.com/bandframework/Taweret}}
 }
 
 @misc{yannotty2023model,
-      title={Model Mixing Using {B}ayesian Additive Regression Trees}, 
+      title={Model mixing using {B}ayesian Additive Regression Trees}, 
       author={John C. Yannotty and Thomas J. Santner and Richard J. Furnstahl and Matthew T. Pratola},
       year={2023},
       journal = {Technometrics},
@@ -116,7 +116,7 @@ title = "{{Taweret: A Python Package for Bayesian Model Mixing}}",
 
 @article{Pan2020,
[…]
