+++ _config.yml
@@ -13,7 +13,7 @@ execute:
 
 # Information about where the book exists on the web
 repository:
-  url: https://https://bandframework.github.io/Taweret/book/index.html  # Online location of your book
+  url: https://bandframework.github.io/Taweret/book/index.html  # Online location of your book
   branch: jupyter-book  # Which branch of the repository should be used when creating links (optional)
 
 # Add GitHub buttons to your book
