+++ .github/workflows/deploy-jupyter-book.yml
@@ -1,51 +1,87 @@
-# Sample workflow for building and deploying a Jekyll site to GitHub Pages
-name: Deploy Taweret Jupyter-Book
+name: deploy-book
 
+# Only run this when the jupyter-book branch changes
 on:
-  # Runs on pushes targeting the default branch
   push:
-    branches: ["jupyter-book"]
+    branches:
+    - jupyter-book
 
-  # Allows you to run this workflow manually from the Actions tab
-  workflow_dispatch:
-
-# Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
-permissions:
-  contents: read
-  pages: write
-  id-token: write
-
-# Allow only one concurrent deployment, skipping runs queued between the run in-progress and latest queued.
-# However, do NOT cancel in-progress runs as we want to allow these production deployments to complete.
-concurrency:
-  group: "pages"
-  cancel-in-progress: false
-
+# This job installs dependencies, build the book, and pushes it to `gh-pages`
 jobs:
-  # Build job
-  build:
+  deploy-book:
     runs-on: ubuntu-latest
     steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-      - name: Setup Pages
-        uses: actions/configure-pages@v3
-      - name: Build with Jekyll
-        uses: actions/jekyll-build-pages@v1
-        with:
-          source: ./
-          destination: ./_site
-      - name: Upload artifact
-        uses: actions/upload-pages-artifact@v2
+    - uses: actions/checkout@v2
 
-  # Deployment job
-  deploy:
-    environment:
-      name: github-pages
-      url: ${{ steps.deployment.outputs.page_url }}
-    runs-on: ubuntu-latest
-    needs: build
-    steps:
-      - name: Deploy to GitHub Pages
-        id: deployment
-        uses: actions/deploy-pages@v2
+    # Install dependencies
+    - name: Set up Python 3.7
+      uses: actions/setup-python@v1
+      with:
+        python-version: 3.7
+
+    # - name: Install dependencies
+    #   run: |
+    #     pip install -r requirements.txt
+    # Build the book
+    - name: Build the book
+      run: |
+        jupyter-book build taweret-book/
+    # Push the book's HTML to github-pages
+    - name: GitHub Pages action
+      uses: peaceiris/actions-gh-pages@v3.7.3
+      with:
+        github_token: ${{ secrets.GITHUB_TOKEN }}
+        publish_dir: taweret-book/_build/html
+        cname: https://bandframework.github.io/Taweret/book/index.html
+
+# # Sample workflow for building and deploying a Jekyll site to GitHub Pages
+# name: Deploy Taweret Jupyter-Book
+
+# on:
+#   # Runs on pushes targeting the default branch
+#   push:
+#     branches: ["jupyter-book"]
+
+#   # Allows you to run this workflow manually from the Actions tab
+#   workflow_dispatch:
+
+# # Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
+# permissions:
+#   contents: read
+#   pages: write
+#   id-token: write
+
+# # Allow only one concurrent deployment, skipping runs queued between the run in-progress and latest queued.
+# # However, do NOT cancel in-progress runs as we want to allow these production deployments to complete.
+# concurrency:
+#   group: "pages"
+#   cancel-in-progress: false
+
+# jobs:
+#   # Build job
+#   build:
+#     runs-on: ubuntu-latest
+#     steps:
+#       - name: Checkout
+#         uses: actions/checkout@v3
+#       - name: Setup Pages
+#         uses: actions/configure-pages@v3
+#       - name: Build with Jekyll
[…]
