### 2024-05-23 #6175 @ominusliticus
Final release version for JOSS publication is: v1.0.2
The zenodo archive is DOI: 10.5281/zenodo.11254359
Authors, ORCIDs, and licenses have been checked
Files are currently uploading and I will publish zenodo record once finalized

Regarding the _length_ of the article, we plan to produce a jupyter-book which will be a slightly expanded version of this paper with the notebooks currently in our documentation page moved there and made interactive.
