### 2024-05-23 #6175 @rkurchin
(note to @openjournals/pe-eics to check [this comment](https://github.com/openjournals/joss-reviews/issues/6175#issuecomment-2111411113) re missing DOI's)
