### 2024-05-24 #6175 @danielskatz
No, it will be the track editor who will do that.
