### 2024-05-24 dbb0e6a
attempts to fix Jupyter Book
