### 2024-05-24 35786fc
deletes _config.yml from outer repo
