### 2024-05-24 #6175 @kyleniemeyer
@openjournals/dev It looks like the `<code>` tag around the software name in the paper was translated to the webpage for the paper: https://joss.theoj.org/papers/10.21105/joss.06175

Does that require a manual fix? Thanks!
