### 2024-05-24 991d381
deletes book.md file
