+++ book/_config.yml
@@ -13,7 +13,7 @@ execute:
 
 # Information about where the book exists on the web
 repository:
-  url: https://github.com/bandframework/Taweret
+  url: https://github.com/bandframework/Taweret/tutorial-book
   path_to_book: book  # Optional path to your book, relative to the repository root
   branch: jupyter-book  # Which branch of the repository should be used when creating links (optional)
 
