+++ docs/run_to_rebuild_tawret_rst.sh
@@ -1,5 +1,5 @@
 cd source/
-sphinx-apidoc -f -E --implicit-namespaces -l -o . ../Taweret
+sphinx-apidoc -f -E --implicit-namespaces -l -o . ../../src/Taweret
 cd ..
 make clean html
 open build/html/Taweret.html
