+++ _sources/index.rst.txt
@@ -1,0 +1,55 @@
+.. Taweret documentation master file, created by
+   sphinx-quickstart on Wed Sep 21 14:40:23 2022.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to Bayesian Model Mixing with Taweret!
+==============================================
+
+.. toctree::
+   :maxdepth: 4
+   :caption: About Taweret:
+
+   intro
+
+.. toctree::
+   :maxdepth: 4
+   :caption: Installation:
+
+   installation
+   macos_user
+
+.. toctree::
+   :maxdepth: 4
+   :caption: API Documentation:
+
+   modules
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Examples:
+   
+   notebooks/Linear_BMM_with_cdf_function_for_coleman_models
+   notebooks/coleman_models_BMM_comparative_study
+   notebooks/Linear_BMM_with_step_function_for_SAMBA_models
+   notebooks/Linear_BMM_with_cdf_function_for_SAMBA_models
+   notebooks/Linear_BMM_with_switchcos_function_for_SAMBA_models
+   notebooks/Bivariate_Test
+   notebooks/N_Model_BMM
+   notebooks/Trees_BMM_2D.ipynb
+   notebooks/Trees_BMM_2D_K_models.ipynb
+
+.. toctree::
+   :maxdepth: 2
+   :caption: For Developers:
+
+   tox_environments
+   how_to_add_new_mixing
+   how_to_add_documentation
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
