+++ _sources/Taweret.core.rst.txt
@@ -1,0 +1,31 @@
+Taweret.core namespace
+======================
+
+.. py:module:: Taweret.core
+
+Submodules
+----------
+
+
+.. automodule:: Taweret.core.base_mixer
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+.. automodule:: Taweret.core.base_model
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+.. automodule:: Taweret.core.trees_setup
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+.. automodule:: Taweret.core.wrappers
+   :members:
+   :undoc-members:
+   :show-inheritance:
