+++ _sources/Taweret.models.rst.txt
@@ -1,0 +1,25 @@
+Taweret.models namespace
+========================
+
+.. py:module:: Taweret.models
+
+Submodules
+----------
+
+
+.. automodule:: Taweret.models.coleman_models
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+.. automodule:: Taweret.models.polynomial_models
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+.. automodule:: Taweret.models.samba_models
+   :members:
+   :undoc-members:
+   :show-inheritance:
