+++ book/intro/bmm_basics.md
@@ -1,0 +1,1 @@
+## Bayesian model mixing: a crash course
