### 2024-05-30 870b90b
(Issue #91) Try building sphinx docs with tox.
I've temporarily set this to run on PRs to develop so that I can test it.
Also, I'm not having it publish any pages for now.  I'll do that once all the
docs are clean.
