### 2024-05-30 fa71fdd
(Issue #91) Archive HTML-format docs.
