+++ src/Taweret/models/samba_models.py
@@ -6,7 +6,8 @@
 # ##########################################################
 
 # imports
-from Taweret.utils.utils import normal_log_likelihood_elementwise as log_likelihood_elementwise_utils
+from Taweret.utils.utils import normal_log_likelihood_elementwise \
+    as log_likelihood_elementwise_utils
 from Taweret.core.base_model import BaseModel
 import numpy as np
 from scipy import special, integrate
@@ -86,7 +87,8 @@ class Loworder(BaseModel):
 
                     if k % 2 == 0:
                         low_c[k] = np.sqrt(
-                            2.0) * special.gamma(k + 0.5) * (-4.0)**(k // 2) / (math.factorial(k // 2))
+                            2.0) * special.gamma(k + 0.5) * (-4.0)**(k // 2) \
+                                / (math.factorial(k // 2))
                     else:
                         low_c[k] = 0
 
@@ -129,8 +131,9 @@ class Loworder(BaseModel):
                 for k in range(int(self.order + 2)):
 
                     if k % 2 == 0:
-                        c[k] = np.sqrt(2.0) * special.gamma(k + 0.5) * (-4.0)**(
-                            k // 2) / (math.factorial(k) * math.factorial(k // 2))
+                        c[k] = np.sqrt(2.0) * special.gamma(k + 0.5) *\
+                              (-4.0)**(k // 2) / (math.factorial(k) *
+                                                  math.factorial(k // 2))
                     else:
                         c[k] = 0.0
 
@@ -140,7 +143,8 @@ class Loworder(BaseModel):
 
                 # variance
                 var1 = (cbar)**2.0 * (math.factorial(self.order + 2)
-                                      )**2.0 * self.x**(2.0 * (self.order + 2))
+                                      )**2.0 * self.x**(2.0 *
+                                                        (self.order + 2))
 
             # model 2 for even orders
             elif self.error_model == 'informative':
@@ -153,8 +157,11 @@ class Loworder(BaseModel):
                         if k == 0:
                             c[k] = 0.0
                         else:
-                            c[k] = np.sqrt(2.0) * special.gamma(k + 0.5) * (-4.0)**(k // 2) / (
-                                math.factorial(k // 2) * math.factorial(k // 2 - 1) * 4.0**(k))
+                            c[k] = (np.sqrt(2.0) * special.gamma(k + 0.5) *
+                                    (-4.0)**(k // 2)
+                                    / (math.factorial(k // 2)
+                                       * math.factorial(k // 2 - 1)
+                                       * 4.0**(k)))
                     else:
                         c[k] = 0.0
 
@@ -163,8 +170,9 @@ class Loworder(BaseModel):
                                (self.order // 2 + 1))
 
                 # variance
-                var1 = (cbar)**2.0 * (math.factorial(self.order // 2)
-                                      )**2.0 * (4.0 * self.x)**(2.0 * (self.order + 2))
+                var1 = (cbar)**2.0 * \
+                    (math.factorial(self.order // 2))**2.0 * \
+                    (4.0 * self.x)**(2.0 * (self.order + 2))
 
         # odd order
         else:
@@ -178,8 +186,9 @@ class Loworder(BaseModel):
                 for k in range(int(self.order + 1)):
 
                     if k % 2 == 0:
-                        c[k] = np.sqrt(2.0) * special.gamma(k + 0.5) * (-4.0)**(
-                            k // 2) / (math.factorial(k) * math.factorial(k // 2))
+                        c[k] = np.sqrt(2.0) * special.gamma(k + 0.5) * \
+                            (-4.0)**(k // 2) \
[…]
