### 2024-06-02 51ece00
deleting book from this branch
