+++ src/Taweret/models/polynomial_models.py
@@ -40,18 +40,20 @@ class polynomal_model(BaseModel):
             Returns:
             --------
             :returns: mean and standard deviation of the
-            model at the x grid points.
+                model at the x grid points.
             :rtype: np.ndarray, np.ndarray
             :return values: mean predictions.
             :return values: standard deviation of the predictions.
 
         '''
+
         if isinstance(x, list):
             x = np.array(x)
         m = self.c * (x - self.a)**self.p + self.b
         if len(m.shape) == 1:
             m = m.reshape(m.shape[0], 1)
         s = np.array([1] * x.shape[0]).reshape(m.shape[0], 1)
+
         return m, s
 
     def set_prior(self):
@@ -99,12 +101,13 @@ class sin_exp(BaseModel):
         Returns:
         --------
         :returns: mean and standard deviation of the model
-        at the x grid points.
+            at the x grid points.
         :rtype: np.ndarray, np.ndarray
         :return values: mean predictions.
         :return values: standard deviation of the predictions.
 
         '''
+
         # Check type of x
         if isinstance(x, list):
             x = np.array(x)
@@ -146,6 +149,7 @@ class sin_exp(BaseModel):
             ts = ts.reshape(ts.shape[0], 1)
 
         s = np.array([1] * x.shape[0]).reshape(ts.shape[0], 1)
+
         return ts, s
 
     def set_prior(self):
@@ -191,12 +195,13 @@ class cos_exp(BaseModel):
         Returns:
         --------
         :returns: mean and standard deviation of the model at the
-        x grid points.
+            x grid points.
         :rtype: np.ndarray, np.ndarray
         :return values: mean predictions.
         :return values: standard deviation of the predictions.
 
         '''
+
         # Check type of x
         if isinstance(x, list):
             x = np.array(x)
@@ -238,6 +243,7 @@ class cos_exp(BaseModel):
             ts = ts.reshape(ts.shape[0], 1)
 
         s = np.array([1] * x.shape[0]).reshape(ts.shape[0], 1)
+
         return ts, s
 
     def set_prior(self):
@@ -293,12 +299,13 @@ class sin_cos_exp(BaseModel):
         Returns:
         --------
         :returns: mean and standard deviation of the model at the x
-        grid points.
+            grid points.
         :rtype: np.ndarray, np.ndarray
         :return values: mean predictions.
         :return values: standard deviation of the predictions.
 
         '''
+
         # Check type of x
         if isinstance(x, list):
             x = np.array(x)
@@ -378,6 +385,7 @@ class sin_cos_exp(BaseModel):
             ts = ts.reshape(ts.shape[0], 1)
 
         s = np.array([1] * x.shape[0]).reshape(ts.shape[0], 1)
+
         return ts, s
 
     def set_prior(self):
