### 2024-06-04 #6175 @kyleniemeyer
@openjournals/dev could we get some help removing the `<code>` tags around the paper's title?
