### 2024-06-05 47e777c
(Issue #91) Update sphinx builds so that notebooks are rendered.
