+++ tox.ini
@@ -48,6 +48,8 @@ commands =
 [testenv:html]
 description = Generate Taweret's HTML documentation
 deps =
+    ipython
+    Pygments
     sphinx
     sphinxcontrib-bibtex
     sphinx_rtd_theme
