### 2024-06-05 71a3316
(Issue #91) See if we can get CI docs actions to use pandoc with notebooks.
