+++ .github/workflows/make_docs.yml
@@ -26,6 +26,10 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
+      - name: Install dependencies
+        run: |
+          sudo apt-get update
+          sudo apt-get -y install pandoc
       - name: Setup base Python environment
         run: |
           $CLONE_PATH/.github/workflows/setup_base_python.sh ${{ runner.os }}
