### 2024-06-06 #6175 @kyleniemeyer
Paper title fixed! https://joss.theoj.org/papers/10.21105/joss.06175
